% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_loinc_system_col.R
\name{add_loinc_system_col}
\alias{add_loinc_system_col}
\title{Get all Loinc System (Specimen) Types for a Lab}
\usage{
add_loinc_system_col(dataframe, concept_id_col = NULL)
}
\arguments{
\item{dataframe}{input data}

\item{concept_id_col}{The column in dataframe that points to the concept_id. If NULL, defaults to "concept_id".}
}
\description{
This function takes a dataframe and mutates an additional column providing the specimen type based on the "Has system" relationship id.
}
