% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_chemidplus_synonyms.R
\name{annotate_chemidplus_synonyms}
\alias{annotate_chemidplus_synonyms}
\title{Get any synonyms in chemidplus}
\usage{
annotate_chemidplus_synonyms(phrase, type = "contains")
}
\description{
Helpful with investigational drugs
}
