% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_atc_classification.R
\name{atc_classification}
\alias{atc_classification}
\title{Vuew ATC Class Descendants}
\usage{
atc_classification(
  conn,
  concept_obj,
  vocab_schema = "omop_vocabulary",
  verbose = TRUE,
  render_sql = TRUE,
  sleep_time = 1
)
}
\description{
Due to the high amount of records, used to determine the appropriate range based on row counts per level to supply the optional `range` argument for \code{\link{plot_loinc_classification}}, where this fucntion is called again and can be optionally filtered on a numeric range before plotting.
}
