% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_query_valueset.R
\name{bulk_query_valueset}
\alias{bulk_query_valueset}
\title{Query a vector of permissible values in bulk}
\usage{
bulk_query_valueset(valueset, type = "exact")
}
\arguments{
\item{valueset}{vector of permissible values to search}

\item{type}{the type of query to perform. Defaults to "exact".}
}
\description{
This function queries the CONCEPT and CONCEPT_SYNONYM table for a vector of permissible values, or valueset,
}
