% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_join.R
\name{cache_join}
\alias{cache_join}
\title{Cache the results of a query}
\usage{
cache_join(
  object,
  function_name,
  left_vector,
  right_table_name,
  right_column_name,
  ...
)
}
\arguments{
\item{object}{object to cache}

\item{key}{equivalent to the SQL query string.}
}
\value{
Invisibly returns the path to the cache file
}
\description{
This function uses the R.cache::saveCache function with the destination directory by Package name "Athena".
}
