% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cacheLeftJoin.R
\name{cacheLeftJoin}
\alias{cacheLeftJoin}
\title{Cache the results of a query}
\usage{
cacheLeftJoin(
  object,
  vector,
  athena_table,
  athena_column,
  where_athena_col,
  where_athena_col_equals,
  omop,
  omop_schema
)
}
\arguments{
\item{object}{object to cache}

\item{key}{equivalent to the SQL query string.}
}
\value{
Invisibly returns the path to the cache file
}
\description{
This function uses the R.cache::saveCache function with the destination directory by Package name "Athena".
}
