% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concept.tree.R
\name{concept.tree}
\alias{concept.tree}
\title{Combine Parents and Children Relationships}
\usage{
concept.tree(
  concept_id,
  parental_generations = 1,
  child_generations = 1,
  override_cache = FALSE,
  verbose = FALSE,
  cache_resultset = TRUE,
  conn = NULL,
  render_sql = TRUE,
  schema,
  style
)
}
\description{
Combine Parents and Children Relationships
}
