% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concept2_triplet.R
\name{concept2_triplet}
\alias{concept2_triplet}
\title{Get a triplet from the concept_relationship table}
\usage{
concept2_triplet(
  .data,
  .col = NULL,
  names_from,
  include_count = TRUE,
  omop = FALSE,
  omop_schema = "omop_vocabulary"
)
}
\arguments{
\item{.data}{data the contains the concept_id (subject)}

\item{.col}{column in data that contains the concept_id (subject)}

\item{names_from}{predicate of the triplet}
}
\description{
Get a triplet from the concept_relationship table
}
