% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concept_parent_functions.R
\name{concept_children}
\alias{concept_children}
\title{Query concept children}
\usage{
concept_children(parent_id, generations = 1, override_cache = FALSE)
}
\description{
Query concept children
}
