% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logical.R
\name{concept_id_exists}
\alias{concept_id_exists}
\title{QA to make sure the concept_id exists in OMOP}
\usage{
concept_id_exists(
  concept_id,
  vocab_schema,
  conn = NULL,
  cache_only = FALSE,
  skip_cache = FALSE,
  override_cache = FALSE,
  render_sql = FALSE,
  verbose = FALSE,
  sleepTime = 1
)
}
\description{
QA to make sure the concept_id exists in OMOP
}
