% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_on_concept_parent_functions.R
\name{concept_parents}
\alias{concept_parents}
\title{Query concept parents}
\usage{
concept_parents(child_id, generations = 1, override_cache = FALSE)
}
\description{
Query concept parents
}
