% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concept_relationships.R
\name{concept_relationships}
\alias{concept_relationships}
\title{Look up the relationships between concept groups and domains}
\usage{
concept_relationships(relationship_id)
}
\description{
Look up the relationships between concept groups and domains
}
