% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concept_id_exists.R
\name{conceptIdExists}
\alias{conceptIdExists}
\title{QA to make sure the concept_id exists in OMOP}
\usage{
conceptIdExists(concept_id, schema)
}
\description{
QA to make sure the concept_id exists in OMOP
}
