% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x_conn.R
\name{connectAthena}
\alias{connectAthena}
\title{Connect Local Athena}
\usage{
connectAthena(
  port = 5432,
  dbname = "athena",
  pathToDriver = system.file(package = "pg13", "driver"),
  verbose = FALSE
)
}
\description{
Connect Local Athena
}
