% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conn.R
\name{connectAthena_ff}
\alias{connectAthena_ff}
\title{Connect Local Athena Function Factory}
\usage{
connectAthena_ff(
  user = NULL,
  password = NULL,
  port = 5432,
  server = "localhost/athena"
)
}
\description{
Connect Local Athena Function Factory
}
