% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.R
\name{constraints}
\alias{constraints}
\title{FUNCTION_TITLE}
\usage{
constraints(
  conn,
  conn_fun = "connectAthena()",
  vocab_schema,
  table,
  field,
  skip_cache = FALSE,
  override_cache = FALSE,
  render_sql = TRUE,
  verbose = TRUE,
  sleepTime = 1
)
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
\seealso{
\code{\link[SqlRender]{render}}
}
