% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_new_vocabulary.R
\name{copy_new_vocabulary}
\alias{copy_new_vocabulary}
\title{Copy New Vocabulary to Athena}
\usage{
copy_new_vocabulary(
 
    folder_path = "~/Desktop/vocabulary_download_v5_{a6b9f6b7-f8ef-4d04-ac57-6cdb7f17fc31}_1593572935845"
)
}
\description{
Copy New Vocabulary to Athena
}
