% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createConceptParentTable.R
\name{createConceptParentTable}
\alias{createConceptParentTable}
\title{Create a CONCEPT_PARENT Table}
\usage{
createConceptParentTable(schema)
}
\description{
This function creates a subset of the CONCEPT_ANCESTOR Table where either max_ or min_ levels of separation are 0 or 1. This table is created to visualize the hierarchy of concepts using external R packages such as collapsibleTrees.
}
