% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSettings.R
\name{createSettings}
\alias{createSettings}
\title{Create a .Settings Object}
\usage{
createSettings()
}
\arguments{
\item{concept_code}{PARAM_DESCRIPTION, Default: NULL}

\item{vocabulary_id}{PARAM_DESCRIPTION, Default: NULL}

\item{domain_id}{PARAM_DESCRIPTION, Default: NULL}

\item{concept_class_id}{PARAM_DESCRIPTION, Default: NULL}

\item{standard_concept}{PARAM_DESCRIPTION, Default: NULL}

\item{invalid_reason}{PARAM_DESCRIPTION, Default: NULL}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
This function takes the Concept Table fields present in the global environment and aggregates them into a list object `.Settings` that can then be run against a data frame as a whole using the \code{\link{filterSettings}} function.
}
\details{
DETAILS
}
\seealso{
\code{\link[rubix]{map_names_set}}
 \code{\link[purrr]{keep}},\code{\link[purrr]{map}}
}
