% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosswalk_from_umls_search.R
\name{crosswalk_from_umls_search}
\alias{crosswalk_from_umls_search}
\title{Crosswalk concept_id to UMLS based on Code}
\usage{
crosswalk_from_umls_search(phrase)
}
\description{
Crosswalk concept_id to UMLS based on Code
}
