% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosswalk_umls_code.R
\name{crosswalk_to_umls_code}
\alias{crosswalk_to_umls_code}
\title{Crosswalk concept_id to UMLS based on Code}
\usage{
crosswalk_to_umls_code(concept_id)
}
\description{
Crosswalk concept_id to UMLS based on Code
}
