% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_family_tree.R
\name{derive_family_tree}
\alias{derive_family_tree}
\title{Make the family tree dataframe required for the ggenealogy package}
\usage{
derive_family_tree(concept_dataframe, concept_dataframe_column = NULL)
}
\arguments{
\item{concept_dataframe}{The concept_dataframe cannot be subsetted since the original concept fields are required to create the labels for the family tree}
}
\value{
dataframe of 1-degree separation labeled as parent and child
}
\description{
Make the family tree dataframe required for the ggenealogy package
}
