% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_loinc_concept_stem.R
\name{derive_loinc_concept_stem}
\alias{derive_loinc_concept_stem}
\title{Get the concept stem of a LOINC concept name}
\usage{
derive_loinc_concept_stem(concept_dataframe)
}
\arguments{
\item{concept_dataframe}{a dataframe pulled from the concept table. Requires the concept_name column at minimum.}
}
\description{
This function allows to isolate the "concept" stem in all concept_names formatted in LOINC's system (ie "{Lab Name [units] Etcetera} where Lab Name is the stem".) This allows to search based on a grepl string match for any concepts that have been missed.
}
