% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_join.R
\name{ds_join_on_ingredient}
\alias{ds_join_on_ingredient}
\title{Get the Staged Drug Strengths by Ingredient}
\usage{
ds_join_on_ingredient(
  data,
  ingredient_concept_id_col = "drug_concept_id",
  conn,
  conn_fun = "connectAthena()",
  ds_schema = "patelm9",
  verbose = TRUE,
  render_sql = TRUE,
  render_only = FALSE
)
}
\arguments{
\item{ds_schema}{Schema for the Drug Strength Staged table, Default: 'patelm9'}
}
\description{
Join the staged calculations in the Drug Strength Staged table by
`ingredient_concept_id`.
}
