% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_lookup_ucum.R
\name{ds_lookup_ucum}
\alias{ds_lookup_ucum}
\title{Lookup the Units for Drug Strength Processing}
\usage{
ds_lookup_ucum(
  conn,
  conn_fun = "connectAthena()",
  vocab_schema = "omop_vocabulary"
)
}
\description{
To normalize the drug exposure calculations in the Drug Strength tables,
all units related to rate, commonly seen in extended release formulations,
or time, are excluded from the final calculation.
}
\seealso{
\code{\link[SqlRender]{render}}

Other drug strength functions: 
\code{\link{ds_process_drug_table}()},
\code{\link{ds_process_map_table}()},
\code{\link{ds_process}()},
\code{\link{ds_stage_unit_fraction}()},
\code{\link{ds_stage_value_fraction}()},
\code{\link{ds_stage}()}
}
\concept{drug strength functions}
