% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_process.R
\name{ds_process}
\alias{ds_process}
\title{Process the Drug Strength Table}
\usage{
ds_process(
  conn,
  conn_fun = "connectAthena()",
  vocab_schema = "omop_vocabulary",
  write_schema = "patelm9",
  verbose = TRUE,
  render_sql = TRUE,
  render_only = FALSE
)
}
\value{
Drug Strength Processed table
}
\description{
The Drug Strength table is filtered for only valid entries and the 'box_size',
'invalid_reason', 'valid_start_date', and 'valid_end_date' fields are selected
against. Concurrently the unit concept ids for the `numerator`, `denominator`, and
`amount` in the table are mapped to their names. The resulting Drug Strength
Processed table is stored in the given schema.

After processing, the values found in the Drug Strength and Drug Strength
Processed tables can be staged for calculations. See `\code{\link{ds_stage}}.
}
\seealso{
\code{\link[SqlRender]{render}}
}
