% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_stage.R
\name{ds_stage_unit_fraction}
\alias{ds_stage_unit_fraction}
\title{Stage the Fractional Representation of Units}
\usage{
ds_stage_unit_fraction(
  conn,
  conn_fun = "connectAthena()",
  write_schema = "patelm9",
  verbose = TRUE,
  render_sql = TRUE,
  render_only = FALSE
)
}
\description{
Stage the Fractional Representation of Units
}
\seealso{
\code{\link[SqlRender]{render}}
}
