% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_join.R
\name{ds_staged_table_exists}
\alias{ds_staged_table_exists}
\title{Does the Drug Strength Staged table exist?}
\usage{
ds_staged_table_exists(
  conn,
  conn_fun = "connectAthena()",
  ds_schema = "patelm9"
)
}
\arguments{
\item{ds_schema}{Schema for the Drug Strength Staged table, Default: 'patelm9'}
}
\description{
Checks for the existence of the Drug Strength Staged table, which is a
prerequisite for any `ds_join_*` functions. If it does not exist, it can be
written using \code{\link{ds_process}} followed by \code{\link{ds_stage}}.
}
\seealso{
\code{\link[rlang]{parse_expr}}
 \code{\link[pg13]{ls_tables}}
}
