% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{extractHemOncRegimens}
\alias{extractHemOncRegimens}
\title{Extract HemOnc Regimens by Component String Match}
\usage{
extractHemOncRegimens(
  component,
  component_count = NULL,
  omop = FALSE,
  omop_schema = "omop_vocabulary"
)
}
\arguments{
\item{component_count}{If NULL or the component_count is larger than the maximum number of components possible for all Regimens with a positive string match to the `component` parameter, the unfiltered result of the initial query for `component` is returned.}
}
\description{
This function uses the grepl function on the HemOnc Concept Names and naming patterns to return HemOnc Regimens based on a single Component and total number of Components in the source regimen.
}
