% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_athena.R
\name{fetched_via}
\alias{fetched_via}
\title{FUNCTION_TITLE}
\usage{
fetched_via(chariotResultsetObj)
}
\arguments{
\item{chariotResultsetObj}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
