% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_for_hemonc.R
\name{filter_for_hemonc}
\alias{filter_for_hemonc}
\title{Returns all possible OMOP domains in Athena}
\usage{
filter_for_hemonc(dataframe)
}
\value{
dataframe of unique `domain_id` variables from the `concept` table
}
\description{
Returns all possible OMOP domains in Athena
}
