% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_for_rxnorm.R
\name{filter_for_rxnorm}
\alias{filter_for_rxnorm}
\title{Returns all possible OMOP domains in Athena}
\usage{
filter_for_rxnorm(dataframe, rxnorm_extension = TRUE)
}
\value{
dataframe of unique `domain_id` variables from the `concept` table
}
\description{
Returns all possible OMOP domains in Athena
}
