% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_invalid_reason.R
\name{filter_invalid_reason}
\alias{filter_invalid_reason}
\title{Filter the invalid_reason Field}
\usage{
filter_invalid_reason(.data, value, invert = FALSE)
}
\description{
Filter the invalid_reason Field
}
