% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_standard_concept.R
\name{filter_standard_concept}
\alias{filter_standard_concept}
\title{Filter the standard_concept field}
\usage{
filter_standard_concept(.data, value, invert = FALSE)
}
\description{
Filter the standard_concept field
}
