% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_functions.R
\name{filterAtStrip}
\alias{filterAtStrip}
\title{Filter Multiple Concept Strip Columns}
\usage{
filterAtStrip(.data, merge_cols, all = TRUE, ...)
}
\arguments{
\item{.data}{Dataframe}

\item{merge_cols}{Character vector of column names to filter}

\item{all}{Equivalent to the `all_vars()` variable predicate in the Tidyverse system. If TRUE, all `merge_cols` are filtered for. If FALSE, the equivalent to `any_vars()` is performed. Default: TRUE}

\item{...}{Filter arguments passed to the dplyr filter function using the base Concept Table fields.}
}
\value{
A tibble.
}
\description{
This function performs the same style of filtering as \code{\link{filterStrip}} over multiple columns.
}
\seealso{
\code{\link{filterStrip}}
\code{\link[dplyr]{select}},\code{\link[dplyr]{select_all}},\code{\link[dplyr]{bind}},\code{\link[dplyr]{filter_all}},\code{\link[dplyr]{filter}},\code{\link[dplyr]{distinct}}
 \code{\link[tidyr]{separate_rows}}
 \code{\link[rubix]{normalize_all_to_na}}
}
