% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_functions.R
\name{filterClassConcepts}
\alias{filterClassConcepts}
\title{Filter Class Concepts}
\usage{
filterClassConcepts(
  .data,
  has_prefix = NULL,
  has_suffix = NULL,
  invert = FALSE
)
}
\description{
Filter Class Concepts
}
