% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_functions.R
\name{filterConcept}
\alias{filterConcept}
\title{Primitive filtering function}
\usage{
filterConcept(
  data,
  has_prefix = NULL,
  has_suffix = NULL,
  concept_col,
  values,
  invert = FALSE
)
}
\arguments{
\item{values}{character vector of length 1 or greater.}
}
\description{
Primitive filtering function
}
\seealso{
\code{\link[dplyr]{filter_all}}
}
