% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_functions.R
\name{filterConceptClass}
\alias{filterConceptClass}
\title{Filter Concept Class}
\usage{
filterConceptClass(
  .data,
  has_prefix = NULL,
  has_suffix = NULL,
  values,
  invert = FALSE
)
}
\arguments{
\item{values}{character vector of length 1 or greater.}
}
\description{
Filter Concept Class
}
\seealso{
\code{\link[dplyr]{filter_all}}
}
