% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_on_concept_parent_functions.R
\name{filterParentlessConcepts}
\alias{filterParentlessConcepts}
\title{Get Topmost Ancestors}
\usage{
filterParentlessConcepts(.data, column = NULL, ...)
}
\arguments{
\item{.data}{data}

\item{...}{Arguments passed to the dplyr filter function}
}
\description{
This function asks the question: of all the concept ids provided in the input data, which ones do not have a parent?. If the question is: of all the concept ids in the input, what are the topmost ancestor concepts?, try \code{\link{topmostAncestors}}.
}
