% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_functions.R
\name{filterRxNorm}
\alias{filterRxNorm}
\title{Filter for RxNorm Concepts}
\usage{
filterRxNorm(
  .data,
  has_prefix = NULL,
  has_suffix = NULL,
  includeExt = TRUE,
  invert = FALSE
)
}
\description{
Filter for RxNorm Concepts
}
