% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_parameters_functions.R
\name{filterSettings}
\alias{filterSettings}
\title{Filter Settings}
\usage{
filterSettings(.data)
}
\arguments{
\item{.data}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{

}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
\seealso{
\code{\link[purrr]{map_if}},\code{\link[purrr]{map2}}
 \code{\link[cave]{vector_to_string}}
 \code{\link[secretary]{typewrite_bold}},\code{\link[secretary]{typewrite}}
 \code{\link[rlang]{parse_expr}}
}
