% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterStandardConcepts.R
\name{filterStandardConcepts}
\alias{filterStandardConcepts}
\title{Filter Standard Concepts}
\usage{
filterStandardConcepts(
  .data,
  has_prefix = NULL,
  has_suffix = NULL,
  invert = FALSE
)
}
\description{
Filter Standard Concepts
}
