% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterStandardConceptType.R
\name{filterStandardConceptType}
\alias{filterStandardConceptType}
\title{Filter by Standard Concept Type}
\usage{
filterStandardConceptType(
  .data,
  has_prefix = NULL,
  has_suffix = NULL,
  values,
  invert = FALSE
)
}
\description{
Filter by Standard Concept Type
}
