% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterStrip.R
\name{filterStrip}
\alias{filterStrip}
\title{Filter Merged Concepts}
\usage{
filterStrip(.data, merge_col, ...)
}
\arguments{
\item{.data}{dataframe that contains the merged concept column}

\item{merge_col}{column of merged concepts}

\item{...}{arguments for filter function using concept table fields}
}
\description{
Filter Merged Concepts
}
