% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filterValid}
\alias{filterValid}
\title{Filter for Valid Concepts}
\usage{
filterValid(.data, has_prefix = NULL, has_suffix = NULL, rm_date_fields = TRUE)
}
\value{
A data frame filtered for `invalid_reason == NA` with or without the `valid_start_date` and `valid_end_date` removed.
}
\description{
Filter a data frame queried from the Concept Table for only the valid concepts based on the `invalid_reason` column with the option to remove the `valid_start_date` and `valid_end_date` columns.
}
\seealso{
\code{\link[dplyr]{filter}},\code{\link[dplyr]{select}}
}
