% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filterVocabulary}
\alias{filterVocabulary}
\title{Filter Vocabulary}
\usage{
filterVocabulary(
  .data,
  has_prefix = NULL,
  has_suffix = NULL,
  values,
  invert = FALSE
)
}
\description{
Filter Vocabulary
}
