% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{get_ancestor_relationships}
\alias{get_ancestor_relationships}
\title{Get Ancestry Relationships}
\usage{
get_ancestor_relationships(
  vocab_schema,
  conn = NULL,
  conn_fun = "connectAthena()",
  cache_only = FALSE,
  skip_cache = FALSE,
  override_cache = FALSE,
  cache_resultset = TRUE,
  render_sql = TRUE,
  verbose = TRUE,
  sleepTime = 1
)
}
\description{
Get Ancestry Relationships
}
\seealso{
\code{\link[SqlRender]{render}}
}
