% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_athena_triples.R
\name{get_athena_triples}
\alias{get_athena_triples}
\title{Get RDF-Style Triples from the Athena vocabulary}
\usage{
get_athena_triples(vocabulary_id = "RxNorm", limit = 1000)
}
\description{
Get RDF-Style Triples from the Athena vocabulary
}
