% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x_lookup.R
\name{get_concept}
\alias{get_concept}
\title{Lookup a Concept Id}
\usage{
get_concept(
  concept_id,
  vocab_schema = "omop_vocabulary",
  conn,
  conn_fun = "connectAthena()",
  cache_only = FALSE,
  skip_cache = FALSE,
  override_cache = FALSE,
  render_sql = FALSE,
  verbose = FALSE,
  sleepTime = 1
)
}
\value{
Concept Class object if the `concept_id` argument is found in the Concept table.
If it is not found, nothing is returned.
}
\description{
Lookup a Concept Id
}
\examples{
library(tidyverse)
library(chariot)

lookup_atc_class_ingredients(atc_concept_obj = 21601787)

lisinopril_class <- get_concept(concept_id = 21601787)
lisinopril_class

lookup_atc_class_ingredients(atc_concept_obj = lisinopril_class)
}
