% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{get_concept_id_fields}
\alias{get_concept_id_fields}
\title{Get the Concept Id Field Names for Each Table}
\usage{
get_concept_id_fields()
}
\description{
Get the Concept Id Field Names for Each Table
}
\seealso{
\code{\link[rubix]{filter_at_grepl}},\code{\link[rubix]{split_deselect}}
 \code{\link[dplyr]{select}}
 \code{\link[purrr]{map}}
}
