% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated2.R
\name{get_conn_db}
\alias{get_conn_db}
\title{Get Connection Database Name}
\usage{
get_conn_db(conn)
}
\description{
This is to make sure that the cache path directory has the same name as the database to prevent collisions between multiple OMOP Vocabulary sources used simulataneously.
}
