% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.R
\name{get_constraints}
\alias{get_constraints}
\title{Get Constraints}
\usage{
get_constraints(
  conn,
  conn_fun = "connectAthena()",
  vocab_schema,
  table,
  field,
  skip_cache = FALSE,
  override_cache = FALSE,
  render_sql = TRUE,
  verbose = TRUE,
  sleepTime = 1
)
}
\value{
A named list of unique values by given field.
}
\description{
For any table in the given schema, a
named list of distinct values within the provided field
is returned.
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
\seealso{
\code{\link[SqlRender]{render}}
}
