% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_name_cluster.R
\name{get_name_cluster}
\alias{get_name_cluster}
\title{Get a Name Cluster}
\usage{
get_name_cluster(
  concept_obj,
  vocab_schema = "omop_vocabulary",
  conn,
  conn_fun = "connectAthena()",
  cache_only = FALSE,
  skip_cache = FALSE,
  override_cache = FALSE,
  render_sql = TRUE,
  verbose = TRUE,
  sleepTime = 1
)
}
\arguments{
\item{concept_obj}{May be an integer or a `concept` class object.}
}
\description{
A `Name Cluster` is a character vector of all the unique labels associated
with a concept, including labels derived from `Maps to` relationships from
the Concept Relationship table.
}
\examples{
library(chariot)
library(tidyverse)

conn <- connectAthena()


# Drug Domain
concept_obj <- get_concept(conn = conn,
                           concept_id = 1308216)
concept_obj
get_name_cluster(conn = conn,
                 concept_obj = concept_obj)


# Cancer Condition Domain
concept_obj <- get_concept(conn = conn,
                           concept_id = 4187868)
concept_obj
get_name_cluster(conn = conn,
                 concept_obj = concept_obj)

# Chronic Disease Domain
concept_obj <- get_concept(conn = conn,
                           concept_id = 319835)
concept_obj
get_name_cluster(conn = conn,
                 concept_obj = concept_obj)

# Measurement Domain
concept_obj <- get_concept(conn = conn, concept_id = 4298431)
concept_obj
get_name_cluster(conn = conn,
                 concept_obj = concept_obj)


dcAthena(conn = conn)
}
