% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x_get_test.R
\name{get_test_data}
\alias{get_test_data}
\title{Get Test Data}
\usage{
get_test_data(
  conn,
  conn_fun = "connectAthena()",
  fields = "*",
  schema = "omop_vocabulary",
  table = "concept",
  limit = 100,
  verbose = TRUE,
  render_sql = TRUE
)
}
\description{
Get test data for further experimentation for functions in this or other
packages.
}
\seealso{
\code{\link[rlang]{parse_expr}}
 \code{\link[SqlRender]{render}}
}
