% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version_key.R
\name{get_VersionKey}
\alias{get_VersionKey}
\title{Get Version Key}
\usage{
get_VersionKey(
  conn,
  conn_fun = "pg13::local_connect(verbose=FALSE)",
  schema = "omop_vocabulary",
  log_schema = "public",
  log_table = "setup_athena_log",
  log_timestamp_field = "sa_datetime",
  template_only = FALSE,
  sql_only = FALSE
)
}
\arguments{
\item{conn}{PARAM_DESCRIPTION}

\item{conn_fun}{PARAM_DESCRIPTION, Default: 'pg13::local_connect(verbose=FALSE)'}

\item{schema}{PARAM_DESCRIPTION, Default: 'omop_vocabulary'}

\item{log_schema}{PARAM_DESCRIPTION, Default: 'public'}

\item{log_table}{PARAM_DESCRIPTION, Default: 'setup_athena_log'}

\item{log_timestamp_field}{PARAM_DESCRIPTION, Default: 'sa_datetime'}

\item{template_only}{PARAM_DESCRIPTION, Default: FALSE}

\item{sql_only}{PARAM_DESCRIPTION, Default: FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
The key used as a hash for the caching feature.
}
\details{
If a `log table` is present as written using the `setupAthena` package,
the VersionKey class object is returned using the last row. If the
`log_table` is not found, the VersionKey is created from a list of the
vocabulary_id-vocabulary_version key-value pair from the Vocabulary table.
}
