% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{get_vocab_data_dictionary}
\alias{get_vocab_data_dictionary}
\title{Get the Vocabulary Data Dictionary}
\usage{
get_vocab_data_dictionary()
}
\description{
Get the Vocabulary Data Dictionary
}
\seealso{
\code{\link[dplyr]{bind}}
 \code{\link[rubix]{format_colnames}}
}
