% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{get_vocabulary_ids}
\alias{get_vocabulary_ids}
\title{Get Vocabulary Ids}
\usage{
get_vocabulary_ids(
  vocab_schema = "omop_vocabulary",
  conn,
  conn_fun = "connectAthena()",
  cache_only = FALSE,
  skip_cache = FALSE,
  override_cache = FALSE,
  cache_resultset = TRUE,
  render_sql = TRUE,
  verbose = TRUE,
  sleepTime = 1
)
}
\description{
Get Vocabulary Ids
}
\seealso{
\code{\link[SqlRender]{render}}
}
