% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getConceptClasses.R, R/getDomain.R
\name{getConceptClasses}
\alias{getConceptClasses}
\alias{getDomain}
\title{Get all OMOP Concept Classes}
\usage{
getConceptClasses(
  schema = NULL,
  verbose = FALSE,
  cache_resultset = TRUE,
  override_cache = FALSE,
  conn = NULL,
  render_sql = FALSE,
  sleepTime = 1,
  ...
)

getDomain(
  schema = NULL,
  verbose = FALSE,
  cache_resultset = TRUE,
  override_cache = FALSE,
  conn = NULL,
  render_sql = FALSE,
  sleepTime = 1,
  ...
)
}
\value{
A data frame of unique `domain_id`, `vocabulary_id`, and `concept_class_id` combinations from the Concept Table.

A data frame of unique `domain_id`, `vocabulary_id`, and `concept_class_id` combinations from the Concept Table.
}
\description{
Get all OMOP Concept Classes

Get all OMOP Domain
}
\seealso{
\code{\link[pg13]{buildQuery}}

\code{\link[pg13]{buildQuery}}
}
