% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_functions.R
\name{getStrip}
\alias{getStrip}
\title{=
Get Strip}
\usage{
getStrip(concept_id, schema = "public")
}
\description{
=
Get Strip
}
\examples{
library(tidyverse)
library(chariot)

getStrip(concept_id = 1112807,
         schema = "omop_vocabulary")


test_data <-
    tibble::tibble(Concept = "[V] [S] 1112807 aspirin [RxNorm 1191] [Drug] [Ingredient]\n[V] [S] 1112807 aspirin [RxNorm 1191] [Drug] [Ingredient]")

unboxStrip(data = test_data,
           strip_col = Concept)
unboxStrip(data = test_data,
           strip_col = Concept,
           remove = TRUE)
}
\seealso{
\code{\link[dplyr]{select}}
}
