% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hemonc_regimens.R
\name{hemonc_regimens}
\alias{hemonc_regimens}
\title{Query all HemOnc regimens with more than 2 components}
\usage{
hemonc_regimens(
  component,
  component_count = NULL,
  omop = FALSE,
  omop_schema = "omop_vocabulary"
)
}
\description{
Query all HemOnc regimens with more than 2 components
}
