% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hemonc.R
\name{ho_lookup_antineoplastics}
\alias{ho_lookup_antineoplastics}
\title{Query a HemOnc Regimen's 'Has antineoplastic' Relationship}
\usage{
ho_lookup_antineoplastics(
  regimen_concept_ids,
  vocab_schema = NULL,
  check_validity = TRUE,
  conn,
  cache_only = FALSE,
  skip_cache = FALSE,
  override_cache = FALSE,
  render_sql = FALSE,
  verbose = FALSE,
  sleepTime = 1
)
}
\description{
Query a HemOnc Regimen's 'Has antineoplastic' Relationship
}
\seealso{
\code{\link[cli]{cli_rule}},\code{\link[cli]{cat_line}}
 \code{\link[SqlRender]{render}}
}
