% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ho_harmonize.R
\name{ho_reduce_to_components}
\alias{ho_reduce_to_components}
\title{Normalize To HemOnc Components}
\usage{
ho_reduce_to_components(..., check_validity = TRUE, schema = "omop_vocabulary")
}
\arguments{
\item{hemonc_concept_ids}{HemOnc Vocabulary Concept Ids of either Regimen or Component concept classes.}
}
\description{
This function takes a mixture of HemOnc Regimen and HemOnc Component Concepts and returns all the unique HemOnc Components associated with the input combination.
}
\seealso{
\code{\link[dplyr]{filter}},\code{\link[dplyr]{select}}
}
