% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_functions.R
\name{innerJoin}
\alias{innerJoin}
\title{INNER JOIN an OMOP Vocabulary Table}
\usage{
innerJoin(
  data,
  column = NULL,
  athena_schema,
  athena_table,
  athena_column,
  where_athena_col = NULL,
  where_athena_col_in = NULL,
  render_sql = TRUE,
  conn = NULL
)
}
\description{
This function executes the join() function with joinType == "INNER".
}
