% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_phrase_query_cached.R
\name{is_phrase_query_cached}
\alias{is_phrase_query_cached}
\title{Is the query_phrase cached?}
\usage{
is_phrase_query_cached(
  phrase,
  type,
  limit = NULL,
  case_insensitive = TRUE,
  where_col = NULL,
  where_col_in = NULL
)
}
\description{
Is the query_phrase cached?
}
