% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x_sql.R
\name{join}
\alias{join}
\title{Join a R dataframe to Athena's concept table}
\usage{
join(
  kind = c("LEFT", "RIGHT", "INNER", "FULL"),
  data,
  column = NULL,
  vocab_table,
  vocab_field,
  select_data_columns = "*",
  select_vocab_fields = "*",
  distinct = FALSE,
  write_schema = "patelm9",
  vocab_schema = "omop_vocabulary",
  where_in_vocab_field,
  where_in_vocab_field_value,
  where_not_in_vocab_field,
  where_not_in_vocab_field_value,
  where_is_null_vocab_field,
  where_is_not_null_vocab_field,
  case_insensitive = TRUE,
  conn,
  conn_fun = "connectAthena()",
  verbose = TRUE,
  render_sql = TRUE,
  render_only = FALSE
)
}
\arguments{
\item{data}{dataframe to join}

\item{column}{string of the column name to join on. If NULL, the 1st column is used.}

\item{vocab_table}{OMOP Vocabulary table to join to}

\item{select_data_columns}{Vector of field names in the `data` to select for.}

\item{select_vocab_fields}{Vector of field names in the `vocab_table` to select for.}

\item{distinct}{Will the resultset contain only distinct rows?}

\item{write_schema}{Schema in the database with write permissions to write the
table containing `data` to. This table is dropped on exit.}

\item{vocab_schema}{Schema containing the OMOP Vocabulary tables.}

\item{vocab_column}{Field in the `vocab_table` to join on.}
}
\description{
This function has better performance than the WHERE IN statement for larger searches. A table in the format of "v{unpunctuated timestamp}" is written to the local Athena. A join is performed with a concept table. The table is subsequently dropped. If a dataframe column is not provided as a the column to join the concept table on, the 1st column will be used by default.
}
