% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprec-join.R, R/join_for.R
\name{join_for_concept_synonym_name}
\alias{join_for_concept_synonym_name}
\title{Join For Concept Synonyms}
\usage{
join_for_concept_synonym_name(
  kind = c("LEFT", "RIGHT", "INNER", "FULL"),
  data,
  concept_id_column = NULL,
  select_data_columns = "*",
  select_concept_synonym_fields = c("concept_id", "concept_synonym_name"),
  distinct = FALSE,
  write_schema = "patelm9",
  vocab_schema = "omop_vocabulary",
  where_in_concept_synonym_field = "language_concept_id",
  where_in_concept_synonym_field_value = 4180186,
  where_not_in_concept_synonym_field,
  where_not_in_concept_synonym_field_value,
  where_is_null_concept_synonym_field,
  where_is_not_null_concept_synonym_field,
  case_insensitive = TRUE,
  conn,
  conn_fun = "connectAthena()",
  verbose = TRUE,
  render_sql = TRUE,
  render_only = FALSE
)

join_for_concept_synonym_name(
  kind = c("LEFT", "RIGHT", "INNER", "FULL"),
  data,
  concept_id_column = NULL,
  select_data_columns = "*",
  select_concept_synonym_fields = c("concept_id", "concept_synonym_name"),
  distinct = FALSE,
  write_schema = "patelm9",
  vocab_schema = "omop_vocabulary",
  where_in_concept_synonym_field = "language_concept_id",
  where_in_concept_synonym_field_value = 4180186,
  where_not_in_concept_synonym_field,
  where_not_in_concept_synonym_field_value,
  where_is_null_concept_synonym_field,
  where_is_not_null_concept_synonym_field,
  case_insensitive = TRUE,
  conn,
  conn_fun = "connectAthena()",
  verbose = TRUE,
  render_sql = TRUE,
  render_only = FALSE
)
}
\description{
`join_for_*` functions differ from `join_on_*` functions in that `join_for_*`
joins on a vocabulary table field that is already specified, and that the join
is to add a specific field to the data, this case being the
`concept_synonym_name` field with a join on the `concept_id` field.

`join_for_*` functions differ from `join_on_*` functions in that `join_for_*`
joins on a vocabulary table field that is already specified, and that the join
is to add a specific field to the data, this case being the
`concept_synonym_name` field with a join on the `concept_id` field.
}
