% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_on.R
\name{join_on_concept_name}
\alias{join_on_concept_name}
\title{FUNCTION_TITLE}
\usage{
join_on_concept_name(
  kind = c("LEFT", "RIGHT", "INNER", "FULL"),
  data,
  column = NULL,
  select_data_columns = "*",
  distinct = FALSE,
  write_schema = "patelm9",
  vocab_schema = "omop_vocabulary",
  where_in_concept_field,
  where_in_concept_field_value,
  where_not_in_concept_field,
  where_not_in_concept_field_value,
  where_is_null_concept_field = "invalid_reason",
  where_is_not_null_concept_field,
  case_insensitive = TRUE,
  conn,
  conn_fun = "connectAthena()",
  verbose = TRUE,
  render_sql = TRUE,
  render_only = FALSE
)
}
\arguments{
\item{kind}{PARAM_DESCRIPTION, Default: c("LEFT", "RIGHT", "INNER", "FULL")}

\item{data}{PARAM_DESCRIPTION}

\item{column}{PARAM_DESCRIPTION, Default: NULL}

\item{select_data_columns}{PARAM_DESCRIPTION, Default: '*'}

\item{distinct}{PARAM_DESCRIPTION, Default: FALSE}

\item{write_schema}{PARAM_DESCRIPTION, Default: 'patelm9'}

\item{vocab_schema}{PARAM_DESCRIPTION, Default: 'omop_vocabulary'}

\item{where_in_concept_field}{PARAM_DESCRIPTION}

\item{where_in_concept_field_value}{PARAM_DESCRIPTION}

\item{where_not_in_concept_field}{PARAM_DESCRIPTION}

\item{where_not_in_concept_field_value}{PARAM_DESCRIPTION}

\item{where_is_null_concept_field}{PARAM_DESCRIPTION, Default: 'invalid_reason'}

\item{where_is_not_null_concept_field}{PARAM_DESCRIPTION}

\item{conn}{PARAM_DESCRIPTION}

\item{conn_fun}{PARAM_DESCRIPTION, Default: 'connectAthena()'}

\item{verbose}{PARAM_DESCRIPTION, Default: TRUE}

\item{render_sql}{PARAM_DESCRIPTION, Default: TRUE}

\item{render_only}{PARAM_DESCRIPTION, Default: FALSE}
}
\description{
FUNCTION_DESCRIPTION
}
