% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{label_to_strip}
\alias{label_to_strip}
\title{Concert a Label Column to a Merge Column}
\usage{
label_to_strip(data, label_col, into_strip_col, remove = FALSE)
}
\description{
Concert a Label Column to a Merge Column
}
\seealso{
\code{\link[tidyr]{extract}}
}
