% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelToMerge.R
\name{labelToMerge}
\alias{labelToMerge}
\title{Concert a Label Column to a Merge Column}
\usage{
labelToMerge(.data, label_col, into, remove = FALSE)
}
\description{
Concert a Label Column to a Merge Column
}
