% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/left_join_concept.R
\name{left_join_concept}
\alias{left_join_concept}
\title{Join a dataframe object with Athena's concept table}
\usage{
left_join_concept(
  .data,
  .col = NULL,
  concept_column = "concept_id",
  include_synonyms = TRUE
)
}
\arguments{
\item{concept_column}{name of concept_column to join dataframe on. Defaults to concept ID.}

\item{dataframe}{dataframe to join}

\item{dataframe_column}{string of the column name to join on. If NULL, the 1st column is use}
}
\description{
This function has better performance than the WHERE IN statement for larger searches. A table in the format of "v{unpunctuated timestamp}" is written to the local Athena. A join is performed with a concept table. The table is subsequently dropped. If a dataframe column is not provided as a the column to join the concept table on, the 1st column will be used by default.
}
