% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/left_join_descendants.R
\name{left_join_descendants}
\alias{left_join_descendants}
\title{Left Join Ancestors to a dataframe of concept_ids}
\usage{
left_join_descendants(dataframe, dataframe_column = NULL, valid_only = TRUE)
}
\description{
Left Join Ancestors to a dataframe of concept_ids
}
