% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/left_join_for_children.R
\name{left_join_for_children}
\alias{left_join_for_children}
\title{Query Parent-Child Relationships from Concept Ancestor Table}
\usage{
left_join_for_children(.data, .col = NULL, override_cache = FALSE)
}
\description{
This function takes all the ancestors and descendants with a level == 1 relationships (either max or min) for plotting.
}
