% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/left_join_for_descendants.R
\name{left_join_for_descendants}
\alias{left_join_for_descendants}
\title{Left Join Ancestors to a dataframe of concept_ids}
\usage{
left_join_for_descendants(.data, .ancestor_id_column = NULL, level = NULL)
}
\description{
Left Join Ancestors to a dataframe of concept_ids
}
