% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/left_join_for_descendants0.R
\name{left_join_for_descendants0}
\alias{left_join_for_descendants0}
\title{Left Join Ancestors to a dataframe of concept_ids}
\usage{
left_join_for_descendants0(dataframe, ancestor_id_column = NULL, level = NULL)
}
\description{
Left Join Ancestors to a dataframe of concept_ids
}
