% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/left_join_relationship.R
\name{left_join_relationship}
\alias{left_join_relationship}
\title{Join a dataframe object with Athena's concept table}
\usage{
left_join_relationship(
  dataframe,
  dataframe_column = NULL,
  merge_concept_2 = TRUE
)
}
\arguments{
\item{dataframe}{dataframe to join}

\item{dataframe_column}{string of the column name to join on. If NULL, the 1st column is used.}

\item{merge_concept_2}{return the merged concept strip for concept_id_2, or as a straightforward concept table dataframe?}
}
\description{
This function has better performance than the WHERE IN statement for larger searches. A table in the format of "v{unpunctuated timestamp}" is written to the local Athena. A join is performed with a concept table. The table is subsequently dropped. If a dataframe column is not provided as a the column to join the concept table on, the 1st column will be used by default.
}
