% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/left_join_relatives.R
\name{left_join_relatives}
\alias{left_join_relatives}
\title{Get all Ancestors and Descendants of a set of concept ids}
\usage{
left_join_relatives(
  .data,
  id_column = NULL,
  ancestor_level = NULL,
  descendant_level = NULL,
  omop = FALSE,
  omop_schema = "omop_vocabulary"
)
}
\description{
Get all Ancestors and Descendants of a set of concept ids
}
