% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/left_join_synonyms.R
\name{left_join_synonyms}
\alias{left_join_synonyms}
\title{Add synonyms}
\usage{
left_join_synonyms(
  dataframe,
  dataframe_column = NULL,
  include_concept_table = FALSE
)
}
\arguments{
\item{include_concept_table}{Are both the concept table contents and concept synonym table contents expected in the output?}
}
\description{
Identical to the left_join_concept with param include_synonyms set to TRUE
}
