% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leftJoin.R
\name{leftJoin}
\alias{leftJoin}
\title{LEFT JOIN an OMOP Vocabulary Table}
\usage{
leftJoin(
  .data,
  column = NULL,
  athena_schema,
  athena_table,
  athena_column,
  where_athena_col = NULL,
  where_athena_col_in = NULL,
  verbose = FALSE,
  conn = NULL,
  render_sql = FALSE,
  sleepTime = 1,
  ...
)
}
\description{
This function executes the join() function with joinType == "LEFT".
}
