% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leftJoinConcept.R
\name{leftJoinConcept}
\alias{leftJoinConcept}
\title{Join a dataframe object with Athena's concept table}
\usage{
leftJoinConcept(
  .data,
  column = NULL,
  athena_schema = "public",
  concept_column = "concept_id",
  render_sql = TRUE,
  synonyms = FALSE,
  conn = NULL
)
}
\arguments{
\item{concept_column}{name of concept_column to join dataframe on. Defaults to concept ID.}

\item{dataframe}{dataframe to join}

\item{dataframe_column}{string of the column name to join on. If NULL, the 1st column is use}
}
\description{
This function has better performance than the WHERE IN statement for larger searches. A table in the format of "v{unpunctuated timestamp}" is written to the local Athena. A join is performed with a concept table. The table is subsequently dropped. If a dataframe column is not provided as a the column to join the concept table on, the 1st column will be used by default.
}
