% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_functions.R
\name{leftJoinFoChildren}
\alias{leftJoinFoChildren}
\title{LEFT JOIN the Concept Parent Table}
\usage{
leftJoinFoChildren(
  data,
  athena_schema,
  parent_id_column = NULL,
  render_sql = TRUE,
  conn = NULL
)
}
\description{
LEFT JOIN the Concept Parent Table
}
