% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprec-join.R
\name{leftJoinForAncestors}
\alias{leftJoinForAncestors}
\title{Left Join a data frame to the Concept Ancestor Table}
\usage{
leftJoinForAncestors(
  data,
  vocab_schema = "public",
  descendant_id_column = NULL,
  whereLevelIn = NULL,
  whereLevelType = NULL,
  verbose = FALSE,
  conn = NULL,
  render_sql = FALSE,
  sleepTime = 1,
  ...
)
}
\arguments{
\item{data}{PARAM_DESCRIPTION}

\item{vocab_schema}{PARAM_DESCRIPTION, Default: 'public'}

\item{descendant_id_column}{PARAM_DESCRIPTION, Default: NULL}

\item{whereLevelIn}{PARAM_DESCRIPTION, Default: NULL}

\item{whereLevelType}{PARAM_DESCRIPTION, Default: NULL}

\item{verbose}{PARAM_DESCRIPTION, Default: FALSE}

\item{conn}{PARAM_DESCRIPTION, Default: NULL}

\item{render_sql}{PARAM_DESCRIPTION, Default: FALSE}

\item{sleepTime}{PARAM_DESCRIPTION, Default: 1}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Left Join a data frame to the Concept Ancestor Table
}
\details{
DETAILS
}
\seealso{
\code{\link[dplyr]{select}},\code{\link[dplyr]{distinct}},\code{\link[dplyr]{mutate-joins}}
 \code{\link[rubix]{rename_all_with_prefix}}
}
