% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leftJoinForDescendants.R
\name{leftJoinForDescendants}
\alias{leftJoinForDescendants}
\title{Left Join a Dataframe to the Concept Ancestor Table}
\usage{
leftJoinForDescendants(
  .data,
  athena_schema = "public",
  ancestor_id_column = NULL,
  whereLevelIn = NULL,
  whereLevelType = NULL,
  render_sql = TRUE,
  conn = NULL
)
}
\description{
Left Join a Dataframe to the Concept Ancestor Table
}
