% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprec-join.R
\name{leftJoinRelationship}
\alias{leftJoinRelationship}
\title{Left Join Relationship}
\usage{
leftJoinRelationship(
  data,
  column = NULL,
  vocab_schema = "public",
  render_sql = TRUE,
  conn = NULL
)
}
\description{
Left Join Relationship
}
