% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_functions.R
\name{leftJoinRelatives}
\alias{leftJoinRelatives}
\title{LEFT JOIN All Relatives}
\usage{
leftJoinRelatives(
  data,
  athena_schema = "public",
  id_column = NULL,
  whereLevelIn = NULL,
  whereLevelType = NULL,
  render_sql = TRUE,
  conn = NULL
)
}
\description{
LEFT JOIN All Relatives
}
