% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprec-join.R
\name{leftJoinSynonymNames}
\alias{leftJoinSynonymNames}
\title{FUNCTION_TITLE}
\usage{
leftJoinSynonymNames(
  data,
  column = NULL,
  write_schema = "public",
  verbose = FALSE,
  render_sql = FALSE,
  sleepTime = 1,
  vocabulary_id,
  domain_id,
  concept_class_id,
  standard_concept,
  invalid_reason,
  conn = NULL,
  omop_vocabulary_schema
)
}
\arguments{
\item{data}{PARAM_DESCRIPTION}

\item{column}{PARAM_DESCRIPTION, Default: NULL}

\item{verbose}{PARAM_DESCRIPTION, Default: FALSE}

\item{render_sql}{PARAM_DESCRIPTION, Default: FALSE}

\item{sleepTime}{PARAM_DESCRIPTION, Default: 1}

\item{conn}{PARAM_DESCRIPTION, Default: NULL}

\item{vocab_schema}{PARAM_DESCRIPTION, Default: 'public'}

\item{concept_synonym_column}{PARAM_DESCRIPTION, Default: 'concept_synonm_name'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\seealso{
\code{\link[stampede]{stamp_this}}
 \code{\link[pg13]{writeTable}}
 \code{\link[SqlRender]{render}}
}
