% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{list_cdm_table_names}
\alias{list_cdm_table_names}
\title{List the Common Data Model Tables}
\usage{
list_cdm_table_names()
}
\description{
List the names of the Common Data Model tables without the Vocabulary tables.
}
