% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{list_defines_ancestry}
\alias{list_defines_ancestry}
\title{FUNCTION_TITLE}
\usage{
list_defines_ancestry(
  vocab_schema,
  conn = NULL,
  conn_fun = "connectAthena()",
  cache_only = FALSE,
  skip_cache = FALSE,
  override_cache = FALSE,
  cache_resultset = TRUE,
  render_sql = TRUE,
  verbose = TRUE,
  sleepTime = 1
)
}
\arguments{
\item{conn}{PARAM_DESCRIPTION, Default: NULL}

\item{conn_fun}{PARAM_DESCRIPTION, Default: 'connectAthena()'}

\item{cache_only}{PARAM_DESCRIPTION, Default: FALSE}

\item{skip_cache}{PARAM_DESCRIPTION, Default: FALSE}

\item{override_cache}{PARAM_DESCRIPTION, Default: FALSE}

\item{cache_resultset}{PARAM_DESCRIPTION, Default: TRUE}

\item{render_sql}{PARAM_DESCRIPTION, Default: TRUE}

\item{verbose}{PARAM_DESCRIPTION, Default: TRUE}

\item{sleepTime}{PARAM_DESCRIPTION, Default: 1}

\item{schema}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
\seealso{
\code{\link[SqlRender]{render}}
}
