% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drug_strength.R
\name{list_drug_strength_units}
\alias{list_drug_strength_units}
\title{List OMOP Unit Concepts in the Drug Strength Table}
\usage{
list_drug_strength_units(conn = NULL, vocabSchema = "omop_vocabulary")
}
\arguments{
\item{conn}{PARAM_DESCRIPTION, Default: NULL}

\item{vocabSchema}{PARAM_DESCRIPTION, Default: 'omop_vocabulary'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Unpivot all of the Drug Strength table unit concept id fields (amount_unit_concept_id, numerator_unit_concept_id, denominator_unit_concept_id) to a unit_concept_id_type field and return the Concept Table entry for each unit concept id.
}
\details{
The resultset is not cached
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[SqlRender]{render}}
}
