% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{list_fields}
\alias{list_fields}
\title{List Table Fields}
\usage{
list_fields(vocab_schema, tableName, conn = NULL)
}
\arguments{
\item{tableName}{Vocabulary Table}

\item{conn}{Connection object. If NULL, automatically queries the local Athena database. Default: NULL}

\item{schema}{Schema of the Vocabulary Table}
}
\value{
Vector of field names for the table.
}
\description{
List Table Fields
}
\seealso{
\code{\link[pg13]{lsFields}}
}
