% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{list_table_abbr}
\alias{list_table_abbr}
\title{Get OMOP Table Abbreviations}
\usage{
list_table_abbr()
}
\value{
Named vector of abbreviations for all OMOP Tables generated by \code{\link[base]{abbreviate}}.
}
\description{
Get OMOP Table Abbreviations
}
\seealso{
\code{\link[stringr]{str_replace}}
}
\concept{list functions}
