% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drug_strength.R
\name{list_unit_concepts}
\alias{list_unit_concepts}
\alias{list_time_unit_concepts}
\title{List UCUM Unit Concepts}
\usage{
list_unit_concepts(conn = NULL, vocabSchema = "omop_vocabulary")

list_time_unit_concepts(conn = NULL, vocabSchema = "omop_vocabulary")
}
\arguments{
\item{conn}{PARAM_DESCRIPTION, Default: NULL}

\item{vocabSchema}{PARAM_DESCRIPTION, Default: 'omop_vocabulary'}
}
\value{
OUTPUT_DESCRIPTION

OUTPUT_DESCRIPTION
}
\description{
List all UCUM Unit Concepts from the Concept Table. To list UCUM Units related to the measurement of Time, see \code{\link{list_time_unit_concepts}}. The resultset is never cached.

List all UCUM Unit Concepts from the Concept Table that are related to measurement of Time based on a broad regex match to the phrases "hour", "minute", "second", "day", "month", "week", and "year". To list all UCUM Units, see \code{\link{list_unit_concepts}}. This resultset is never cached.
}
\details{
DETAILS

DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
\seealso{
\code{\link[SqlRender]{render}}

\code{\link[SqlRender]{render}}
}
