% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_athena_query.R
\name{load_athena_query}
\alias{load_athena_query}
\title{Load a cached query}
\usage{
load_athena_query(sql_statement, verbose = FALSE)
}
\arguments{
\item{sql_statement}{SQL statement}
}
\description{
This function differs from the query_athena function in that it only returns the cached query resultset if it exists and does not query the local database.
}
