% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{load_cached_query}
\alias{load_cached_query}
\title{Load cacheed results of a query}
\usage{
load_cached_query(key)
}
\arguments{
\item{key}{equivalent to the SQL query string.}
}
\value{
Cached query resultset
}
\description{
This function uses the R.cache::loadCache function with the destination directory set to Package name "Athena".
}
