% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locate_phrase.R
\name{locate_phrase}
\alias{locate_phrase}
\title{Locates the vocabulary table and search type where the phrase is first found}
\usage{
locate_phrase(phrase)
}
\description{
Useful if a simple search isn't returning results.
}
