% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locators.R
\name{locatePhrase}
\alias{locatePhrase}
\title{Locates the vocabulary table and search type where the phrase is first found}
\usage{
locatePhrase(phrase, split, schema)
}
\description{
Useful if a simple search isn't returning results.
}
