% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{loincSystems}
\alias{loincSystems}
\title{LOINC Systems}
\format{
A vector of 13 values:
\describe{
  \item{\code{loincSystems}}{integer Uncommon LOINC Systems}
}
}
\usage{
loincSystems
}
\description{
All the LOINC Systems in the OMOP Vocabulary that are rarely clinically used.
}
\details{
DETAILS
}
\keyword{datasets}
