% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x_lookup_atc_class_ingredients.R
\name{lookup_atc_class_ingredients}
\alias{lookup_atc_class_ingredients}
\title{Lookup the RxNorm Ingredients of an ATC Class}
\usage{
lookup_atc_class_ingredients(
  conn,
  conn_fun = "connectAthena()",
  vocab_schema = "omop_vocabulary",
  atc_concept_obj,
  cache_only = FALSE,
  skip_cache = FALSE,
  override_cache = FALSE,
  render_sql = FALSE,
  verbose = FALSE,
  sleepTime = 1
)
}
\description{
List all the RxNorm Ingredients that belong to an ATC Class.
}
\examples{
library(tidyverse)
library(chariot)

lookup_atc_class_ingredients(atc_concept_obj = 21601787)

lisinopril_class <- get_concept(concept_id = 21601787)
lisinopril_class

lookup_atc_class_ingredients(atc_concept_obj = lisinopril_class)
}
\seealso{
\code{\link[cli]{cli_h1}},\code{\link[cli]{cli_alert}}
 \code{\link[SqlRender]{render}}
}
