% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_concept_ancestry.R
\name{lookup_cid_parent}
\alias{lookup_cid_parent}
\title{Lookup the Parents of a Concept Id}
\usage{
lookup_cid_parent(
  concept_ids,
  vocabSchema,
  conn = NULL,
  skip_cache = FALSE,
  cache_only = FALSE,
  override_cache = FALSE,
  verbose = FALSE,
  render_sql = FALSE,
  sleepTime = 1
)
}
\arguments{
\item{vocabSchema}{PARAM_DESCRIPTION}

\item{conn}{PARAM_DESCRIPTION, Default: NULL}

\item{skip_cache}{PARAM_DESCRIPTION, Default: FALSE}

\item{cache_only}{PARAM_DESCRIPTION, Default: FALSE}

\item{override_cache}{PARAM_DESCRIPTION, Default: FALSE}

\item{verbose}{PARAM_DESCRIPTION, Default: FALSE}

\item{render_sql}{PARAM_DESCRIPTION, Default: FALSE}

\item{sleepTime}{PARAM_DESCRIPTION, Default: 1}

\item{vocabulary_id}{PARAM_DESCRIPTION}

\item{concept_class_id}{PARAM_DESCRIPTION}

\item{domain_id}{PARAM_DESCRIPTION}
}
\value{
Concepts from the Concept Table that fit the provided arguments and that do not have an ancestor in the Concept Ancestor Table.
}
\description{
While \code{\link{lookup_orphan_concepts}} looks up the orphan concepts for an entire concept set, a similar operation can be performed on any given vector of concept_ids. In \code{\link{lookup_orphan_concepts}}, the returned value is a dataframe from the Concept Table of all the concepts that are orphaned in the concept set. Here, the concept_id is looked up regardless with a field in position 1 that is NA if the concept_id is an orphan or a concept_id if the concept_id does have a parent in the Concept Ancestor Table.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[SqlRender]{render}}
}
