% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_concept_ancestry.R
\name{lookup_cid_top_ancestors}
\alias{lookup_cid_top_ancestors}
\title{Lookup Topmost Ancestors of a Concept Id}
\usage{
lookup_cid_top_ancestors(
  vocabSchema,
  concept_ids,
  conn = NULL,
  skip_cache = FALSE,
  cache_only = FALSE,
  override_cache = FALSE,
  verbose = FALSE,
  render_sql = FALSE,
  sleepTime = 1
)
}
\arguments{
\item{vocabSchema}{PARAM_DESCRIPTION}

\item{conn}{PARAM_DESCRIPTION, Default: NULL}

\item{skip_cache}{PARAM_DESCRIPTION, Default: FALSE}

\item{cache_only}{PARAM_DESCRIPTION, Default: FALSE}

\item{override_cache}{PARAM_DESCRIPTION, Default: FALSE}

\item{verbose}{PARAM_DESCRIPTION, Default: FALSE}

\item{render_sql}{PARAM_DESCRIPTION, Default: FALSE}

\item{sleepTime}{PARAM_DESCRIPTION, Default: 1}

\item{vocabulary_id}{PARAM_DESCRIPTION}

\item{concept_class_id}{PARAM_DESCRIPTION}

\item{domain_id}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Trace the topmost lineage of one or more concept ids. To operate on an entire concept set, see \code{\link{lookup_top_ancestors}}.
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
\seealso{
\code{\link[SqlRender]{render}}
}
