% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.R
\name{lookup_concept_id}
\alias{lookup_concept_id}
\title{Lookup a Concept Id}
\usage{
lookup_concept_id(
  concept_id,
  vocab_schema,
  conn = NULL,
  cache_only = FALSE,
  skip_cache = FALSE,
  override_cache = FALSE,
  render_sql = FALSE,
  verbose = FALSE,
  sleepTime = 1
)
}
\description{
Lookup a Concept Id
}
