% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vocabulary_relationships.R
\name{lookup_intervocabulary_relationship}
\alias{lookup_intervocabulary_relationship}
\title{Query a Source Vocabulary's Relationships to Other Vocabularies}
\usage{
lookup_intervocabulary_relationship(
  vocabulary_id,
  conn,
  conn_fun = "connectAthena()",
  vocab_schema = "omop_vocabulary",
  cache_only = FALSE,
  skip_cache = FALSE,
  override_cache = FALSE,
  cache_resultset = TRUE,
  render_sql = FALSE,
  verbose = FALSE,
  sleepTime = 1
)
}
\arguments{
\item{vocabulary_id}{Vector of 1 or more `vocabulary_id` from the OMOP Vocabulary}

\item{conn}{Connection object. If provided, diverts queries to the connection instead of the local Athena instance without caching features.}

\item{cache_only}{Loads from the cache and does not query the database. A NULL object is returned if a resultset was not cached.}

\item{skip_cache}{Skip the caching altogether and directly query the database.}

\item{override_cache}{If TRUE, the cache will not be loaded and will be overwritten by a new query. For override_cache to take effect, skip_cache should be FALSE.}

\item{render_sql}{If TRUE, the SQL will be printed back in the console prior to execution. Default: FALSE}

\item{verbose}{If TRUE, prints loading and querying operations messages to the console. Default: FALSE}

\item{sleepTime}{Argument for `Sys.sleep()` in between queries to allow for halting function execution, especially in cases where other chariot functions are executing multiple queries in succession and require cancellation.}
}
\description{
This family of functions explores the OMOP Vocabulary at the Source Vocabulary level. Lookup functions find details about the given vocabulary while the Query functions query for the inter- and intra-relationships between vocabularies. The concept_class_id for the vocabulary is retained to provide a layer of granularity in the resultsets.
}
\details{
Query for the Source Vocabulary's valid relationships in the Concept Relationship Table and all of the OMOP Source Vocabulary and Concept Classes that the relationships are to. For a resultset for relationships within the vocabulary, see \code{\link{lookup_intravocabulary_relationship}}.
}
\seealso{
\code{\link[SqlRender]{render}}
}
