% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concept_set_ancestry.R
\name{lookup_orphan_concepts}
\alias{lookup_orphan_concepts}
\title{Lookup Orphan Concepts for a Domain, Vocabulary, and/or Concept Class}
\usage{
lookup_orphan_concepts(
  vocabSchema,
  vocabulary_id,
  concept_class_id,
  domain_id,
  conn = NULL,
  skip_cache = FALSE,
  cache_only = FALSE,
  override_cache = FALSE,
  verbose = FALSE,
  render_sql = FALSE,
  sleepTime = 1
)
}
\arguments{
\item{vocabSchema}{PARAM_DESCRIPTION}

\item{vocabulary_id}{PARAM_DESCRIPTION}

\item{concept_class_id}{PARAM_DESCRIPTION}

\item{domain_id}{PARAM_DESCRIPTION}

\item{conn}{PARAM_DESCRIPTION, Default: NULL}

\item{skip_cache}{PARAM_DESCRIPTION, Default: FALSE}

\item{cache_only}{PARAM_DESCRIPTION, Default: FALSE}

\item{override_cache}{PARAM_DESCRIPTION, Default: FALSE}

\item{verbose}{PARAM_DESCRIPTION, Default: FALSE}

\item{render_sql}{PARAM_DESCRIPTION, Default: FALSE}

\item{sleepTime}{PARAM_DESCRIPTION, Default: 1}
}
\value{
Concepts from the Concept Table that fit the provided arguments and that do not have an ancestor in the Concept Ancestor Table.
}
\description{
Returns input concepts filtered by Domain, Vocabulary, and/or Concept Class that do not have valid ancestors.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[SqlRender]{render}}
}
