% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_relationship.R
\name{lookup_relationships}
\alias{lookup_relationships}
\title{Lookup Relationships of a Set of Concepts}
\usage{
lookup_relationships(
  ...,
  check_validity = TRUE,
  conn,
  conn_fun = "connectAthena()",
  vocab_schema = "omop_vocabulary",
  cache_only = FALSE,
  skip_cache = FALSE,
  override_cache = FALSE,
  render_sql = FALSE,
  verbose = FALSE,
  sleepTime = 1
)
}
\arguments{
\item{...}{Concept ids.}

\item{check_validity}{Check to ensure thet the concept ids provided exist}
}
\description{
Returns the forward relationship with concept attributes
of both sides of the relationship.
}
\seealso{
\code{\link[cli]{cli_rule}}
 \code{\link[rlang]{list2}}
 \code{\link[SqlRender]{render}}
}
