% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.R
\name{lookup_synonyms}
\alias{lookup_synonyms}
\title{Lookup Synonyms}
\usage{
lookup_synonyms(
  concept_id,
  vocab_schema,
  conn = NULL,
  cache_only = FALSE,
  skip_cache = FALSE,
  override_cache = FALSE,
  render_sql = FALSE,
  verbose = FALSE,
  sleepTime = 1
)
}
\description{
Lookup Synonyms
}
