% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{lookup_top_classes}
\alias{lookup_top_classes}
\title{Get the Top Classes}
\usage{
lookup_top_classes(
  vocabulary_id,
  domain_id,
  vocab_schema = "omop_vocabulary",
  conn,
  conn_fun = "connectAthena()",
  cache_only = FALSE,
  skip_cache = FALSE,
  override_cache = FALSE,
  cache_resultset = TRUE,
  render_sql = TRUE,
  verbose = TRUE,
  sleepTime = 1
)
}
\description{
Get the top classes in the Concept Ancestor table by vocabulary with the option of filtering further for domain.
}
