% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_units.R
\name{lookup_ucum}
\alias{lookup_ucum}
\title{Lookup UCUM Unit Concepts}
\usage{
lookup_ucum(conn, conn_fun, vocab_schema = "omop_vocabulary")
}
\description{
List all UCUM Unit Concepts from the Concept Table. To list UCUM Units specifically
related to the measurement of Time and Rate, see \code{\link{lookup_ucum_time}}.
To list the UCUM Units used to process the Drug Strength table, see
\code{\link{ds_lookup_ucum}}. The resultset is never cached.
}
\seealso{
\code{\link[SqlRender]{render}}
}
