% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_units.R
\name{lookup_ucum_time}
\alias{lookup_ucum_time}
\title{List UCUM Concepts Related to Time}
\usage{
lookup_ucum_time(conn, conn_fun, vocab_schema = "omop_vocabulary")
}
\description{
List all UCUM Unit Concepts from the Concept Table that are related to
measurement of Time based on a broad regex match to the phrases "hour",
"minute", "second", "day", "month", "week", and "year". To list all UCUM
Units, see \code{\link{lookup_ucum}}. To list the UCUM Units used to process
the Drug Strength table, see \code{\link{ds_lookup_ucum}}This resultset is
never cached.
}
\seealso{
\code{\link[SqlRender]{render}}
}
