% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search-utils.R
\name{make_where_clause}
\alias{make_where_clause}
\title{FUNCTION_TITLE}
\usage{
make_where_clause(
  vocabulary_id,
  domain_id,
  concept_class_id,
  standard_concept,
  invalid_reason
)
}
\arguments{
\item{vocabulary_id}{PARAM_DESCRIPTION}

\item{domain_id}{PARAM_DESCRIPTION}

\item{concept_class_id}{PARAM_DESCRIPTION}

\item{standard_concept}{PARAM_DESCRIPTION}

\item{invalid_reason}{PARAM_DESCRIPTION}

\item{vocabSchema}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
\seealso{
\code{\link[SqlRender]{render}}
}
