% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.R
\name{map_concept_id}
\alias{map_concept_id}
\title{Map All The Concepts in a Dataframe}
\usage{
map_concept_id(
  data,
  write_schema = "patelm9",
  vocab_schema = "omop_vocabulary",
  conn,
  conn_fun = "connectAthena()",
  skip_cache = FALSE,
  override_cache = FALSE,
  cache_only = FALSE,
  cache_resultset = TRUE,
  render_sql = TRUE,
  render_only = FALSE,
  verbose = TRUE,
  sleepTime = 1
)
}
\description{
Map All The Concepts in a Dataframe
}
\seealso{
\code{\link[rlang]{parse_expr}}
 \code{\link[stringr]{str_remove}}
 \code{\link[dplyr]{select_all}},\code{\link[dplyr]{vars}},\code{\link[dplyr]{reexports}},\code{\link[dplyr]{filter_all}},\code{\link[dplyr]{all_vars}},\code{\link[dplyr]{mutate-joins}}
}
