% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_omop_concept_elements.R
\name{merge_omop_concept_elements}
\alias{merge_omop_concept_elements}
\title{Merge OMOP concept elements into a single string}
\usage{
merge_omop_concept_elements(concept_dataframe, into, suffix = NULL)
}
\arguments{
\item{concept_dataframe}{output from concept table}

\item{into}{name of the column that the new combined string will be in}

\item{suffix}{if the omop concept element column names are different from the standard by a suffix, include it so it can point to the correct set of columns}
}
\description{
All elements of the CONCEPT table are included except for the dates.
}
