% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeLabel.R, R/mergeStrip.R
\name{mergeStrip}
\alias{mergeStrip}
\title{Merge OMOP Concepts into a Strip}
\usage{
mergeStrip(.data, into, ..., has_suffix = NULL, has_prefix = NULL)

mergeStrip(.data, into, ..., has_suffix = NULL, has_prefix = NULL)
}
\arguments{
\item{.data}{dataframe with the following required fields from the output}

\item{into}{name of the column that the new combined string will be in}

\item{...}{columns other than concept_id that will be removed in tidyr unite but should be preserved in addition to be merged.}

\item{has_suffix}{if the omop concept element column names are different from the standard by a suffix, include it so it can point to the correct set of columns}

\item{has_prefix}{if the omop concept element column names are prefixed, include it so it can point to the correct set of columns}

\item{label}{if TRUE, only have a strip in the format {concept_id} {concept_name}}

\item{keep_other_cols}{TRUE if all the non-concept table columns in the input dataframe is desired in the output}

\item{shorthand}{This only returns the validity, standard, concept_id, and concept_name as a string. Please note that this merge cannot be unmerged using the unmerge_concepts function.}
}
\value{
A tibble with all blank and "NA" normalized to `<NA>` with 1. If present, `valid_start_date` and `valid_end_date` fields are permanently removed, 2. 8 out of the 10 remaining Concept Table fields (concept_id, concept_name, domain_id, vocabulary_id, concept_class_id, standard_concept, concept_code, invalid_reason) are merged into a single column with the provided column name, 3. the concept_id column is renamed to the format of the provided merged column name: {into_}concept_id. The remaining of the 7 Concept Table fields may also be preserved outside of the merge if provided. All other columns present in the input data are returned along with the transformations described.
}
\description{
All elements of the CONCEPT table are included except for the dates. While this function and unmerge_concepts are meant to be the inverse of one another, it is important to note that this function will select for only the columns in the input that match the pattern of {has_prefix}column_name{has_suffix} and therefore may need to be rejoined to the input after completion. The inverse unmerge_concepts will not remove any columns in the output. It uses the tidyr::extract function to operate directly on the provided column.

This function takes a set of the OMOP Vocabulary Concept Table fields and merges all of them except for the date fields into a single Concept "Strip". If the Strip output is `<NA>` while the input concept id is not, a flagMergeStrip object is returned in the Global Environment.
}
