% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{normalizeToHemOncComponents}
\alias{normalizeToHemOncComponents}
\title{Normalize To HemOnc Components}
\usage{
normalizeToHemOncComponents(hemonc_concept_ids, schema = NULL)
}
\arguments{
\item{hemonc_concept_ids}{HemOnc Vocabulary Concept Ids of either Regimen or Component concept classes.}
}
\description{
This function takes a mixture of HemOnc Regimen and HemOnc Component Concepts and returns all the unique HemOnc Components associated with the input combination.
}
