% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phrase_sql.R
\name{phrase_sql}
\alias{phrase_sql}
\title{Construct SQL for query_phrase}
\usage{
phrase_sql(
  phrase,
  type,
  limit = NULL,
  case_insensitive = TRUE,
  where_col = NULL,
  where_col_in = NULL
)
}
\description{
Also serves as the cache key
}
