% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_by_relative_level.R
\name{pivot_by_relative_level}
\alias{pivot_by_relative_level}
\title{Get all Loinc System (Specimen) Types for a Lab}
\usage{
pivot_by_relative_level(
  .data,
  id_col = NULL,
  levels_type = c("min", "max"),
  include_count = TRUE,
  omop = FALSE,
  omop_schema = "omop_vocabulary"
)
}
\arguments{
\item{concept_id_col}{The column in dataframe that points to the concept_id. If NULL, defaults to "concept_id".}

\item{dataframe}{input data}
}
\description{
This function takes a dataframe and mutates an additional column providing the specimen type based on the "Has system" relationship id.
}
\examples{
Random immunosuppressant concept ids
immunosuppressant_concept_ids <- c("35807335","35807331", "21603616", "21600651", "21605199", "21602723") 
}
