% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_concept2.R
\name{pivot_concept2}
\alias{pivot_concept2}
\title{Get all Loinc System (Specimen) Types for a Lab}
\usage{
pivot_concept2(.data, .col = NULL, names_from, include_count = TRUE)
}
\arguments{
\item{concept_id_col}{The column in dataframe that points to the concept_id. If NULL, defaults to "concept_id".}

\item{dataframe}{input data}
}
\description{
This function takes a dataframe and mutates an additional column providing the specimen type based on the "Has system" relationship id.
}
