% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_concept_2.R
\name{pivot_concept_2}
\alias{pivot_concept_2}
\title{Get all Loinc System (Specimen) Types for a Lab}
\usage{
pivot_concept_2(
  dataframe,
  concept_id_col = NULL,
  names_from,
  include_count = TRUE
)
}
\arguments{
\item{dataframe}{input data}

\item{concept_id_col}{The column in dataframe that points to the concept_id. If NULL, defaults to "concept_id".}
}
\description{
This function takes a dataframe and mutates an additional column providing the specimen type based on the "Has system" relationship id.
}
