% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_wider_by_relationship_concept_class.R
\name{pivot_wider_by_relationship_concept_class}
\alias{pivot_wider_by_relationship_concept_class}
\title{Get all Loinc System (Specimen) Types for a Lab}
\usage{
pivot_wider_by_relationship_concept_class(dataframe, concept_id_col = NULL)
}
\arguments{
\item{dataframe}{input data}

\item{concept_id_col}{The column in dataframe that points to the concept_id. If NULL, defaults to "concept_id".}
}
\description{
This function takes a dataframe and mutates an additional column providing the specimen type based on the "Has system" relationship id.
}
