% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_functions.R
\name{pivotConcept2}
\alias{pivotConcept2}
\title{FUNCTION_TITLE}
\usage{
pivotConcept2(
  .data,
  column = NULL,
  athena_schema = "public",
  render_sql = TRUE,
  conn = NULL,
  names_from,
  include_count = TRUE
)
}
\arguments{
\item{.data}{PARAM_DESCRIPTION}

\item{column}{PARAM_DESCRIPTION, Default: NULL}

\item{athena_schema}{PARAM_DESCRIPTION, Default: 'public'}

\item{render_sql}{PARAM_DESCRIPTION, Default: TRUE}

\item{conn}{PARAM_DESCRIPTION, Default: NULL}

\item{names_from}{PARAM_DESCRIPTION}

\item{include_count}{PARAM_DESCRIPTION, Default: TRUE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[tidyr]{pivot_wider}}
 \code{\link[centipede]{no_na}}
 \code{\link[dplyr]{mutate_all}},\code{\link[dplyr]{select_all}},\code{\link[dplyr]{mutate-joins}},\code{\link[dplyr]{rename}}
}
