% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_functions.R
\name{pivotRelationshipId}
\alias{pivotRelationshipId}
\title{Get all Loinc System (Specimen) Types for a Lab}
\usage{
pivotRelationshipId(
  .data,
  column = NULL,
  athena_schema = "public",
  render_sql = TRUE,
  conn = NULL,
  include_count = TRUE
)
}
\arguments{
\item{concept_id_col}{The column in dataframe that points to the concept_id. If NULL, defaults to "concept_id".}

\item{dataframe}{input data}
}
\description{
This function takes a dataframe and mutates an additional column providing the specimen type based on the "Has system" relationship id.
}
