% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_functions.R
\name{pivotRelative}
\alias{pivotRelative}
\title{FUNCTION_TITLE}
\usage{
pivotRelative(
  .data,
  athena_schema = "public",
  column = NULL,
  whereLevelIn = NULL,
  whereLevelType = NULL,
  names_from,
  include_count = TRUE,
  render_sql = TRUE,
  conn = NULL
)
}
\arguments{
\item{.data}{PARAM_DESCRIPTION}

\item{athena_schema}{PARAM_DESCRIPTION, Default: 'public'}

\item{column}{PARAM_DESCRIPTION, Default: NULL}

\item{whereLevelIn}{PARAM_DESCRIPTION, Default: NULL}

\item{whereLevelType}{PARAM_DESCRIPTION, Default: NULL}

\item{names_from}{PARAM_DESCRIPTION}

\item{include_count}{PARAM_DESCRIPTION, Default: TRUE}

\item{render_sql}{PARAM_DESCRIPTION, Default: TRUE}

\item{conn}{PARAM_DESCRIPTION, Default: NULL}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
\seealso{
\code{\link[dplyr]{select}},\code{\link[dplyr]{mutate_all}},\code{\link[dplyr]{select_all}},\code{\link[dplyr]{mutate-joins}}
 \code{\link[tidyr]{pivot_wider}}
 \code{\link[centipede]{no_na}}
}
