% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_atc_classification.R, R/plot_concept.R
\name{plot_atc_classification}
\alias{plot_atc_classification}
\title{Plot a LOINC Class}
\usage{
plot_atc_classification(
  conn,
  concept_obj,
  range,
  file,
  vocab_schema = "omop_vocabulary",
  color_by = "standard_concept",
  verbose = TRUE,
  render_sql = TRUE,
  sleep_time = 1
)

plot_atc_classification(
  conn,
  concept_obj,
  range,
  file,
  vocab_schema = "omop_vocabulary",
  color_by = "standard_concept",
  verbose = TRUE,
  render_sql = TRUE,
  sleep_time = 1
)
}
\arguments{
\item{skip_plot}{If true, returns the dataframe before it is plotted and plotting is not done. This is an option to troubleshoot or customize a plot beyond what is available within the function.}
}
\description{
Plot a LOINC Class. Due to the high amount of records, use \code{\link{loinc_classification}} to determine the appropriate range based on row counts per level to supply the optional `range` argument.

Plot a LOINC Class. Due to the high amount of records, use \code{\link{loinc_classification}} to determine the appropriate range based on row counts per level to supply the optional `range` argument.
}
\seealso{
\code{\link[tibble]{tibble}}
 \code{\link[dplyr]{bind}},\code{\link[dplyr]{mutate_all}},\code{\link[dplyr]{distinct}},\code{\link[dplyr]{group_by}},\code{\link[dplyr]{summarise_all}},\code{\link[dplyr]{vars}},\code{\link[dplyr]{select}},\code{\link[dplyr]{mutate-joins}}
 \code{\link[tidyr]{pivot_longer}},\code{\link[tidyr]{unite}}
 \code{\link[colorspace]{rainbow_hcl}}
 \code{\link[secretary]{c("typewrite", "typewrite")}},\code{\link[secretary]{press_enter}}
 \code{\link[collapsibleTree]{collapsibleTreeNetwork}}

\code{\link[tibble]{tibble}}
 \code{\link[dplyr]{bind}},\code{\link[dplyr]{mutate_all}},\code{\link[dplyr]{distinct}},\code{\link[dplyr]{group_by}},\code{\link[dplyr]{summarise_all}},\code{\link[dplyr]{vars}},\code{\link[dplyr]{select}},\code{\link[dplyr]{mutate-joins}}
 \code{\link[tidyr]{pivot_longer}},\code{\link[tidyr]{unite}}
 \code{\link[colorspace]{rainbow_hcl}}
 \code{\link[secretary]{c("typewrite", "typewrite")}},\code{\link[secretary]{press_enter}}
 \code{\link[collapsibleTree]{collapsibleTreeNetwork}}
}
