% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapsibleTrees_functions.R
\name{plot_classification}
\alias{plot_classification}
\title{FUNCTION_TITLE}
\usage{
plot_classification(
  vocabulary_id,
  domain_id,
  conn,
  vocab_schema,
  range = 1:10,
  color_by = "vocabulary_id",
  terminal_vocabulary_id,
  verbose = TRUE,
  render_sql = TRUE
)
}
\arguments{
\item{vocabulary_id}{PARAM_DESCRIPTION}

\item{domain_id}{PARAM_DESCRIPTION}

\item{conn}{PARAM_DESCRIPTION}

\item{vocab_schema}{PARAM_DESCRIPTION}

\item{range}{PARAM_DESCRIPTION, Default: 1:10}

\item{color_by}{PARAM_DESCRIPTION, Default: 'vocabulary_id'}

\item{terminal_vocabulary_id}{PARAM_DESCRIPTION}

\item{render_sql}{PARAM_DESCRIPTION, Default: TRUE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[tibble]{tibble}}
 \code{\link[SqlRender]{render}}
 \code{\link[dplyr]{mutate}},\code{\link[dplyr]{select}},\code{\link[dplyr]{reexports}},\code{\link[dplyr]{distinct}},\code{\link[dplyr]{bind}},\code{\link[dplyr]{mutate_all}},\code{\link[dplyr]{group_by}},\code{\link[dplyr]{summarise_all}},\code{\link[dplyr]{mutate-joins}}
 \code{\link[tidyr]{unite}},\code{\link[tidyr]{pivot_longer}}
 \code{\link[colorspace]{rainbow_hcl}}
 \code{\link[collapsibleTree]{collapsibleTreeNetwork}}
}
