% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_concept_hierarchy.R
\name{plot_concept_hierarchy}
\alias{plot_concept_hierarchy}
\title{Make the family tree dataframe required for the ggenealogy package}
\usage{
plot_concept_hierarchy(
  concept_id,
  generations_before = 2,
  generations_after = 2
)
}
\arguments{
\item{concept_dataframe}{The concept_dataframe cannot be subsetted since the original concept fields are required to create the labels for the family tree}
}
\value{
dataframe of 1-degree separation labeled as parent and child
}
\description{
Make the family tree dataframe required for the ggenealogy package
}
