% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_family_tree.R
\name{plot_family_tree}
\alias{plot_family_tree}
\title{Plot the ancestry of a concept}
\usage{
plot_family_tree(
  concept_id,
  family_tree_df,
  generations_before,
  generations_after
)
}
\description{
Plot ancestry  out of the dataframe returned from derive_family_tree()
}
