% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plotBottleneck}
\alias{plotBottleneck}
\title{Plot a Bottleneck Style Reingold-Tilford tree diagram}
\usage{
plotBottleneck(
  concept_id,
  parental_generations = 1,
  child_generations = 1,
  override_cache = FALSE,
  verbose = FALSE,
  cache_resultset = TRUE,
  conn = NULL,
  render_sql = TRUE,
  schema
)
}
\description{
A bottleneck style Reingold-Tilford tree diagram is one where a given set of parent and child relationship levels are plotted where they converge on the single input concept_id in the center. This differs from an "umbrella" style Reingold-Tilford tree diagram where the topmost parent is derived and all the children from that point onwards are returned.
}
