% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprec-plot_functions.R
\name{plotUmbrella}
\alias{plotUmbrella}
\title{Plot an Umbrella Style Reingold-Tilford tree diagram}
\usage{
plotUmbrella(
  concept_id,
  parental_generations = 1,
  child_generations = 1,
  override_cache = FALSE,
  verbose = FALSE,
  cache_resultset = TRUE,
  conn = NULL,
  render_sql = TRUE,
  schema
)
}
\description{
An "umbrella" style Reingold-Tilford tree diagram where the topmost parent is derived and all the children from that point onwards are returned. See `plotBottleneck` to learn more about the Bottleneck-style diagram.
}
