% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_concept.R
\name{preview_rxnorm_atc_classification}
\alias{preview_rxnorm_atc_classification}
\title{Preview Concept Lineage}
\usage{
preview_rxnorm_atc_classification(
  conn,
  rxnorm_concept_obj,
  vocab_schema = "omop_vocabulary",
  verbose = TRUE,
  render_sql = TRUE,
  sleep_time = 1
)
}
\description{
Preview Concept Lineage
}
