% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_manipulation.R
\name{prime_concepts}
\alias{prime_concepts}
\title{Make Modifications to Concepts for SQL querying}
\usage{
prime_concepts(data, concept_col)
}
\arguments{
\item{data}{Dataframe}

\item{concept_col}{Pointer to the column with the concepts to map}
}
\description{
A series of modifications are required for concepts to avoid errors when using them as arguments for SQL querying such as the single quote in "Kaposi's sarcoma" requiring a replacement with an additional "'" escape character, resulting in a string that reads "Kaposi''s sarcoma".
}
\seealso{
\code{\link[dplyr]{tidyeval-compat}},\code{\link[dplyr]{mutate_all}}
 \code{\link[stringr]{str_replace}},\code{\link[stringr]{str_remove}}
}
