% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x_print_concept_hierarchy.R
\name{print_concept_hierarchy}
\alias{print_concept_hierarchy}
\title{Print a Concepts Hierarchy to the Console}
\usage{
print_concept_hierarchy(
  concept_obj,
  level_of_separation_type = c("max", "min"),
  write_schema = "patelm9",
  vocab_schema = "omop_vocabulary",
  conn,
  conn_fun = "connectAthena()",
  cache_only = FALSE,
  skip_cache = FALSE,
  override_cache = FALSE,
  verbose = TRUE,
  render_sql = TRUE,
  render_only = FALSE
)
}
\arguments{
\item{concept_obj}{Concept class object or a concept id.}
}
\description{
If a level of the hierarchy has more than 10 concepts, the first 10 are
printed with an ellipses is included to indicate only the first 10 rows are
shown.
}
\examples{
library(chariot)
library(tidyverse)

conn <- connectAthena()
# Drug Domain
concept_obj <- get_concept(conn = conn,
                           concept_id = 1308216)
concept_obj
print_concept_hierarchy(conn = conn,
                        concept_obj = concept_obj)
print_concept_hierarchy(conn = conn,
                        concept_obj = concept_obj,
                        level_of_separation_type = "min")


# Cancer Condition Domain
concept_obj <- get_concept(conn = conn,
                           concept_id = 4187868)
concept_obj
print_concept_hierarchy(conn = conn,
                        concept_obj = concept_obj)
print_concept_hierarchy(conn = conn,
                        concept_obj = concept_obj,
                        level_of_separation_type = "min")

# Chronic Disease Domain
concept_obj <- get_concept(conn = conn,
                           concept_id = 319835)
concept_obj
print_concept_hierarchy(conn = conn,
                        concept_obj = concept_obj)
print_concept_hierarchy(conn = conn,
                        concept_obj = concept_obj,
                        level_of_separation_type = "min")

# Measurement Domain
concept_obj <- get_concept(conn = conn, concept_id = 4298431)
concept_obj
print_concept_hierarchy(conn = conn,
                        concept_obj = concept_obj)
print_concept_hierarchy(conn = conn,
                        concept_obj = concept_obj,
                        level_of_separation_type = "min")
}
\seealso{
\code{\link[secretary]{character(0)}}
 \code{\link[tibble]{tibble}}
 \code{\link[rubix]{split_by}}
 \code{\link[purrr]{map}}
 \code{\link[dplyr]{arrange}}
}
