% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x_print_concept_siblings.R
\name{print_concept_siblings}
\alias{print_concept_siblings}
\title{Print a Concept's Parents and Siblings}
\usage{
print_concept_siblings(
  concept_obj,
  level_of_separation_type = c("max", "min"),
  write_schema = "patelm9",
  vocab_schema = "omop_vocabulary",
  conn,
  conn_fun = "connectAthena()",
  cache_only = FALSE,
  skip_cache = FALSE,
  override_cache = FALSE,
  verbose = TRUE,
  render_sql = TRUE,
  render_only = FALSE
)
}
\arguments{
\item{concept_obj}{Concept class object or a concept id.}
}
\description{
Print all of a concept's parents and siblings.
}
\examples{
library(chariot)
library(tidyverse)

conn <- connectAthena()
# Drug Domain
concept_obj <- get_concept(conn = conn,
                           concept_id = 1308216)
concept_obj
print_concept_siblings(conn = conn,
                        concept_obj = concept_obj)
print_concept_siblings(conn = conn,
                        concept_obj = concept_obj,
                        level_of_separation_type = "min")


# Cancer Condition Domain
concept_obj <- get_concept(conn = conn,
                           concept_id = 4187868)
concept_obj
print_concept_siblings(conn = conn,
                        concept_obj = concept_obj)
print_concept_siblings(conn = conn,
                        concept_obj = concept_obj,
                        level_of_separation_type = "min")

# Chronic Disease Domain
concept_obj <- get_concept(conn = conn,
                           concept_id = 319835)
concept_obj
print_concept_siblings(conn = conn,
                        concept_obj = concept_obj)
print_concept_siblings(conn = conn,
                        concept_obj = concept_obj,
                        level_of_separation_type = "min")

# Measurement Domain
concept_obj <- get_concept(conn = conn, concept_id = 4298431)
concept_obj
print_concept_siblings(conn = conn,
                        concept_obj = concept_obj)
print_concept_siblings(conn = conn,
                        concept_obj = concept_obj,
                        level_of_separation_type = "min")

dcAthena(conn = conn)
}
\seealso{
\code{\link[rlang]{parse_expr}}
 \code{\link[secretary]{character(0)}}
 \code{\link[tibble]{tibble}}
 \code{\link[dplyr]{filter}},\code{\link[dplyr]{select}}
 \code{\link[rubix]{split_deselect}}
 \code{\link[purrr]{map}}
}
