% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_manipulation.R
\name{process_first_parentheses}
\alias{process_first_parentheses}
\title{Process a Concept Column for the First Set of Parentheses}
\usage{
process_first_parentheses(data, concept_col)
}
\arguments{
\item{data}{Dataframe}

\item{concept_col}{Pointer to the column with the concepts to map}
}
\description{
Process a Concept Column for the First Set of Parentheses
}
\details{
This is the same function as \code{\link{process_parentheses}} except that the Concept Column name is appended with "1st Parentheses" instead of "Parentheses" to accommodate a call to   \code{link{process_last_parentheses}}.
}
