% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_functions.R
\name{process_last_parentheses}
\alias{process_last_parentheses}
\title{Process a Concept Column for the Last Set of Parentheses}
\usage{
process_last_parentheses(data, concept_col)
}
\arguments{
\item{data}{Dataframe}

\item{concept_col}{Pointer to the column with the concepts to map}
}
\description{
Process a Concept Column for the Last Set of Parentheses
}
\details{
This function differs from \code{\link{process_parentheses}} and \code{link{process_last_parentheses}} by performing a greedy regex match for parentheses in cases where there are multiple parenthetical statements within the concept column.
}
