% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_functions.R
\name{process_longest_words}
\alias{process_longest_words}
\title{Identify the Longest Word in the Concept}
\usage{
process_longest_words(data, concept_col, words_col)
}
\arguments{
\item{data}{Dataframe}

\item{concept_col}{Pointer to the column with the concepts to map}

\item{words_col}{Words column produced by \code{\link{process_words}}}
}
\description{
Identify the Longest Word in the Concept
}
\seealso{
\code{\link[dplyr]{tidyeval-compat}},\code{\link[dplyr]{mutate-joins}},\code{\link[dplyr]{group_by_all}},\code{\link[dplyr]{mutate}},\code{\link[dplyr]{filter}},\code{\link[dplyr]{select}},\code{\link[dplyr]{rename}}
}
