% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_manipulation.R
\name{process_map_input_functions}
\alias{process_map_input_functions}
\title{Process Incoming Concepts}
\arguments{
\item{data}{Dataframe}

\item{concept_col}{Pointer to the column with the concepts to map}
}
\description{
Process Incoming Concepts
}
\keyword{internal}
