% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_manipulation.R
\name{process_parentheses}
\alias{process_parentheses}
\title{Process a Concept Column for Parenthetical Phrases}
\usage{
process_parentheses(data, concept_col)
}
\arguments{
\item{data}{Dataframe}

\item{concept_col}{Pointer to the column with the concepts to map}
}
\description{
Process a Concept Column for Parenthetical Phrases
}
\details{
To process a Concept Column with multiple embedded parenthetical phrases, see \code{\link{process_first_parentheses}} and \code{\link{process_last_parentheses}}.
}
