% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_functions.R
\name{process_unique_words}
\alias{process_unique_words}
\title{Identify Words Unique to the Concept in the Concept Set}
\usage{
process_unique_words(data, words_col)
}
\arguments{
\item{data}{Dataframe}

\item{words_col}{Words column produced by \code{\link{process_words}}}
}
\description{
Identify Words Unique to the Concept in the Concept Set
}
\seealso{
\code{\link[dplyr]{tidyeval-compat}},\code{\link[dplyr]{mutate-joins}},\code{\link[dplyr]{count}},\code{\link[dplyr]{filter}},\code{\link[dplyr]{mutate}},\code{\link[dplyr]{select}}
}
