% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_functions.R
\name{process_words}
\alias{process_words}
\title{Process the Words Column from the Concept Column}
\usage{
process_words(data, concept_col, sep = " ", word_nchar = 3)
}
\arguments{
\item{data}{Dataframe}

\item{concept_col}{Pointer to the column with the concepts to map}

\item{sep}{Argument passed to the tidyr::separate_rows, Default: ' '}

\item{word_nchar}{Number of theshold characters for words to be over to be included in the output, Default: 3}
}
\description{
Process the Words Column from the Concept Column
}
