% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_2_hemonc_component_regimens.R
\name{query_2_hemonc_component_regs}
\alias{query_2_hemonc_component_regs}
\title{Query all HemOnc regimens with 2 components}
\usage{
query_2_hemonc_component_regs(component)
}
\description{
Query all HemOnc regimens with 2 components
}
