% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_4_hemonc_component_regs.R
\name{query_4_hemonc_component_regs}
\alias{query_4_hemonc_component_regs}
\title{Query HemOnc regimens with 4 components}
\usage{
query_4_hemonc_component_regs(component)
}
\description{
Query HemOnc regimens with 4 components
}
