% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_ancestors.R
\name{query_ancestors}
\alias{query_ancestors}
\title{Query ancestors for a given concept_id}
\usage{
query_ancestors(descendant_concept_id, max_levels_of_separation)
}
\arguments{
\item{descendant_concept_id}{OMOP concept id of the ancestor concept}

\item{max_levels_of_separation}{Integer of the maximum levels of separation}
}
\description{
Query ancestors for a given concept_id
}
