% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_athena.R
\name{query_athena}
\alias{query_athena}
\title{Generic query function}
\usage{
query_athena(sql_statement, verbose = FALSE)
}
\arguments{
\item{sql_statement}{SQL statement}
}
\description{
This is the low level function that is wrapped for more specific queries. The resultset is cached using the R.cache package
}
