% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_code.R
\name{query_code}
\alias{query_code}
\title{Query based on search terms that does not write to catalogue}
\usage{
query_code(code, limit = NULL, type = c("like", "exact"))
}
\arguments{
\item{...}{vector of codes to collectively feed into the LIKE sql statement}
}
\value{
resultset as a dataframe with all column types as character and trimmed white space
}
\description{
Query based on search terms that does not write to catalogue
}
