% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_descendants.R
\name{query_descendants}
\alias{query_descendants}
\title{Query descendants for a given concept_id}
\usage{
query_descendants(
  ancestor_concept_ids,
  max_levels_of_separation = NULL,
  plot = FALSE
)
}
\description{
Query descendants for a given concept_id
}
