% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hemonc_regimens.R
\name{query_hemonc_regimens}
\alias{query_hemonc_regimens}
\title{Query all HemOnc regimens with more than 2 components}
\usage{
query_hemonc_regimens(component)
}
\description{
Query all HemOnc regimens with more than 2 components
}
