% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_id_synonyms.R
\name{query_id_synonyms}
\alias{query_id_synonyms}
\title{Query the synonyms associated with an id}
\usage{
query_id_synonyms(concept_id)
}
\arguments{
\item{concept_id}{omop concept_id}
}
\description{
Query the synonyms associated with an id
}
