% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_maps_to.R
\name{query_maps_to}
\alias{query_maps_to}
\title{Query all "Maps to" relationships of a concept_id}
\usage{
query_maps_to(concept_id)
}
\arguments{
\item{concept_id}{omop concept_id}
}
\description{
Query all "Maps to" relationships of a concept_id
}
