% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_n_hemonc_component_regimens.R
\name{query_n_hemonc_component_regs}
\alias{query_n_hemonc_component_regs}
\title{Query all HemOnc regimens with more than 2 components}
\usage{
query_n_hemonc_component_regs(component)
}
\description{
Query all HemOnc regimens with more than 2 components
}
