% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_phrase.R
\name{query_phrase}
\alias{query_phrase}
\title{Query based on search terms that does not write to catalogue}
\usage{
query_phrase(
  phrase,
  limit = NULL,
  type = c("like", "exact"),
  case_insensitive = TRUE
)
}
\arguments{
\item{...}{vector of phrases to collectively feed into the LIKE sql statement}
}
\value{
resultset as a dataframe with all column types as character and trimmed white space
}
\description{
Query based on search terms that does not write to catalogue
}
