% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_phrase_synonym.R
\name{query_phrase_synonym}
\alias{query_phrase_synonym}
\title{Query based on search terms in the concept synonym table}
\usage{
query_phrase_synonym(
  phrase,
  limit = NULL,
  type = c("like", "exact"),
  case_insensitive = TRUE
)
}
\arguments{
\item{phrase}{vector of phrases to collectively feed into the LIKE sql statement}
}
\value{
resultset as a dataframe with all column types
}
\description{
Query based on search terms in the concept synonym table
}
