% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_phrases.R
\name{query_phrases}
\alias{query_phrases}
\title{Iterate query_phrase function over a vector of phrases}
\usage{
query_phrases(
  phrases,
  type = c("like", "exact"),
  limit = NULL,
  case_insensitive = TRUE,
  set_names_to = NULL,
  return_valid_only = TRUE
)
}
\arguments{
\item{phrases}{vector of phrases to iterate the query_phrase function on}
}
\description{
Iterate query_phrase function over a vector of phrases
}
