% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_relationship_name.R
\name{query_relationship_name}
\alias{query_relationship_name}
\title{Query relationship_id in relationship table}
\usage{
query_relationship_name(phrase, type = c("like", "exact"))
}
\description{
Query relationship_id in relationship table
}
