% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_string_as_vector.R
\name{query_string_as_vector}
\alias{query_string_as_vector}
\title{Query based on search terms that does not write to catalogue}
\usage{
query_string_as_vector(string, split = " ", limit = NULL)
}
\arguments{
\item{...}{vector of phrases to collectively feed into the LIKE sql statement}
}
\value{
resultset as a dataframe with all column types as character and trimmed white space
}
\description{
Query based on search terms that does not write to catalogue
}
