% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_string_as_vector_synonym.R
\name{query_string_as_vector_synonym}
\alias{query_string_as_vector_synonym}
\title{Query based on search terms that does not write to catalogue}
\usage{
query_string_as_vector_synonym(
  string,
  split = " |[[:punct:]]",
  limit = NULL,
  case_insensitive = TRUE,
  verbose = TRUE,
  override_cache = FALSE
)
}
\arguments{
\item{...}{vector of phrases to collectively feed into the LIKE sql statement}
}
\value{
resultset as a dataframe with all column types as character and trimmed white space
}
\description{
Query based on search terms that does not write to catalogue
}
