% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_vocab_relationships.R
\name{query_vocab_relationships}
\alias{query_vocab_relationships}
\title{Query all the relationships for a vocabulary id}
\usage{
query_vocab_relationships(vocabulary_id)
}
\description{
Query all the relationships for a vocabulary id
}
