% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vocabulary_level_functions.R
\name{query_vocabulary_relationships}
\alias{query_vocabulary_relationships}
\title{Query for a Source Vocabulary's Relationships to Other Source Vocabularies}
\usage{
query_vocabulary_relationships(
  vocabulary_id_1,
  vocabulary_id_2 = NULL,
  conn = NULL,
  cache_only = FALSE,
  skip_cache = FALSE,
  override_cache = FALSE,
  cache_resultset = TRUE,
  render_sql = FALSE,
  verbose = FALSE,
  sleepTime = 1
)
}
\arguments{
\item{vocabulary_id_1}{Vector of 1 or more `vocabulary_id`}

\item{vocabulary_id_2}{Vector of 1 or more target `vocabulary_id`. If NULL, the target vocabulary is set to itself. Default: NULL.}

\item{conn}{Connection object. If provided, diverts queries to the connection instead of the local Athena instance without caching features.}

\item{cache_only}{Loads from the cache and does not query the database. A NULL object is returned if a resultset was not cached.}

\item{skip_cache}{Skip the caching altogether and directly query the database.}

\item{override_cache}{If TRUE, the cache will not be loaded and will be overwritten by a new query. For override_cache to take effect, skip_cache should be FALSE.}

\item{cache_resultset}{(deprecated) If TRUE, the resultset from the query will first be loaded from the cache. The query will be executed if a cached resultset is not retrieved for this particular query, after which the resultset will be cached. If FALSE, Athena or conn will be directly queried without any caching operations.}

\item{render_sql}{If TRUE, the SQL will be printed back in the console prior to execution. Default: FALSE}

\item{verbose}{If TRUE, prints loading and querying operations messages to the console. Default: FALSE}

\item{sleepTime}{Argument for `Sys.sleep()` in between queries to allow for halting function execution, especially in cases where other chariot functions are executing multiple queries in succession and require cancellation.}
}
\description{
This family of functions explores the OMOP Vocabulary at the Source Vocabulary level. Lookup functions find details about the given vocabulary while the Query functions query for the inter- and intra-relationships between vocabularies. The concept_class_id for the vocabulary is retained to provide a layer of granularity in the resultsets.
}
\details{
Query for the Source Vocabulary's non-null relationships in the Concept Relationship Table to a second target set of OMOP Source Vocabulary and its Concept Classes. For a resultset that does not filters on both ends of the relationship, see \code{\link{query_all_vocabulary_relationships}}.
}
\seealso{
\code{\link[SqlRender]{render}}
}
