% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_where_in.R
\name{query_where_in}
\alias{query_where_in}
\title{Query based on search terms that does not write to catalogue}
\usage{
query_where_in(column_name, inclusion_vector, limit = NULL)
}
\arguments{
\item{...}{vector of codes to collectively feed into the LIKE sql statement}
}
\value{
resultset as a dataframe with all column types as character and trimmed white space
}
\description{
Query based on search terms that does not write to catalogue
}
