% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryCode.R
\name{queryCode}
\alias{queryCode}
\title{Query based on search terms that does not write to catalogue}
\usage{
queryCode(
  code,
  schema,
  caseInsensitive = TRUE,
  limit = NULL,
  verbose = FALSE,
  cache_resultset = TRUE,
  override_cache = FALSE,
  conn = NULL,
  render_sql = FALSE,
  type = c("like", "exact")
)
}
\arguments{
\item{...}{vector of codes to collectively feed into the LIKE sql statement}
}
\value{
resultset as a dataframe with all column types as character and trimmed white space
}
\description{
Query based on search terms that does not write to catalogue
}
