% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_functions.R
\name{queryConceptChildren}
\alias{queryConceptChildren}
\title{Query concept children}
\usage{
queryConceptChildren(
  parent_id,
  generations = 1,
  schema,
  conn = NULL,
  cache_only = FALSE,
  skip_cache = FALSE,
  override_cache = FALSE,
  render_sql = FALSE,
  verbose = FALSE,
  sleepTime = 1
)
}
\description{
Query concept children
}
