% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R, R/query_functions.R
\name{queryConceptClassRelationships}
\alias{queryConceptClassRelationships}
\title{FUNCTION_TITLE}
\usage{
queryConceptClassRelationships(
  vocabulary_id_1,
  vocabulary_id_2 = NULL,
  schema = NULL,
  conn = NULL,
  cache_only = FALSE,
  skip_cache = FALSE,
  override_cache = FALSE,
  render_sql = FALSE,
  verbose = FALSE,
  sleepTime = 1
)

queryConceptClassRelationships(
  vocabulary_id_1,
  vocabulary_id_2 = NULL,
  schema = NULL,
  conn = NULL,
  cache_only = FALSE,
  skip_cache = FALSE,
  override_cache = FALSE,
  render_sql = FALSE,
  verbose = FALSE,
  sleepTime = 1
)
}
\arguments{
\item{vocabulary_id_1}{single vocabulary_id for the first vocabulary (Subject)}

\item{vocabulary_id_2}{single vocabulary_id for the second vocabulary (Object). If vocabulary_id_2 is NULL, vocabulary_id_2 is set to vocabulary_id_1.}

\item{conn}{Connection object. If provided, diverts queries to the connection instead of the local Athena instance without caching features.}

\item{cache_only}{Loads from the cache and does not query the database. A NULL object is returned if a resultset was not cached.}

\item{skip_cache}{Skip the caching altogether and directly query the database.}

\item{override_cache}{If TRUE, the cache will not be loaded and will be overwritten by a new query. For override_cache to take effect, skip_cache should be FALSE.}

\item{render_sql}{If TRUE, the SQL will be printed back in the console prior to execution. Default: FALSE}

\item{verbose}{If TRUE, prints loading and querying operations messages to the console. Default: FALSE}

\item{sleepTime}{Argument for `Sys.sleep()` in between queries to allow for halting function execution, especially in cases where other chariot functions are executing multiple queries in succession and require cancellation.}
}
\value{
a [tibble][tibble::tibble-package]

A dataframe with 3 columns: 1) concept_class_id of vocabulary_1 with name as "{vocabulary_id_1}_concept_class_id" unless vocabulary_id_2 is NULL, in which case it will be concept_class_id_1 2) relationship_id from the CONCEPT_RELATIONSHIP table, and 3) concept_class_id of vocabulary_2 with name as "{vocabulary_id_2}_concept_class_id" unless vocabulary_id_2 is NULL, in which case it will be concept_class_id_2.
}
\description{
FUNCTION_DESCRIPTION

This function retrieves the Subject-Predicate-Object triplet from the CONCEPT_RELATIONSHIP table between 2 vocabularies.
}
\seealso{
Other query functions: 
\code{\link{executeAthena}()},
\code{\link{queryAncestors}()},
\code{\link{queryCode}()},
\code{\link{queryConceptId}()},
\code{\link{queryDescendants}()},
\code{\link{queryRelationships}()},
\code{\link{querySynonyms}()},
\code{\link{queryVocabularyRelationships}()},
\code{\link{searchPhraseExactSynonym}()},
\code{\link{searchPhraseExact}()},
\code{\link{searchPhraseLikeSynonym}()},
\code{\link{searchPhraseLike}()},
\code{\link{searchPhraseStringSynonym}()},
\code{\link{searchPhraseString}()},
\code{\link{searchPhrase}()}
}
\concept{query functions}
