% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryConceptId.R
\name{queryConceptId}
\alias{queryConceptId}
\title{Lookup a concept id in Athena}
\usage{
queryConceptId(
  concept_ids,
  schema,
  verbose = FALSE,
  cache_resultset = TRUE,
  override_cache = FALSE,
  conn = NULL,
  render_sql = FALSE,
  sleepTime = 1,
  ...
)
}
\description{
Lookup a concept id in Athena
}
