% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_functions.R
\name{queryConceptParent}
\alias{queryConceptParent}
\title{Query concept children}
\usage{
queryConceptParent(
  child_id,
  schema,
  generations = 1,
  conn = NULL,
  cache_only = FALSE,
  skip_cache = FALSE,
  override_cache = FALSE,
  render_sql = FALSE,
  verbose = FALSE,
  sleepTime = 1
)
}
\description{
Query concept children
}
\seealso{
\code{\link[pg13]{buildQuery}}
 \code{\link[dplyr]{filter}},\code{\link[dplyr]{select}}
 \code{\link[purrr]{keep}}
}
