% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-query_functions.R, R/query_functions.R
\name{queryFamilyTree}
\alias{queryFamilyTree}
\title{FUNCTION_TITLE}
\usage{
queryFamilyTree(
  concept_id,
  schema,
  parental_generations = 1,
  child_generations = 1,
  conn = NULL,
  cache_only = FALSE,
  skip_cache = FALSE,
  override_cache = FALSE,
  render_sql = FALSE,
  verbose = FALSE,
  sleepTime = 1
)

queryFamilyTree(
  concept_id,
  schema,
  parental_generations = 1,
  child_generations = 1,
  conn = NULL,
  cache_only = FALSE,
  skip_cache = FALSE,
  override_cache = FALSE,
  render_sql = FALSE,
  verbose = FALSE,
  sleepTime = 1
)
}
\arguments{
\item{concept_id}{PARAM_DESCRIPTION}

\item{schema}{PARAM_DESCRIPTION}

\item{parental_generations}{PARAM_DESCRIPTION, Default: 1}

\item{child_generations}{PARAM_DESCRIPTION, Default: 1}

\item{conn}{PARAM_DESCRIPTION, Default: NULL}

\item{cache_only}{PARAM_DESCRIPTION, Default: FALSE}

\item{skip_cache}{PARAM_DESCRIPTION, Default: FALSE}

\item{override_cache}{PARAM_DESCRIPTION, Default: FALSE}

\item{render_sql}{PARAM_DESCRIPTION, Default: FALSE}

\item{verbose}{PARAM_DESCRIPTION, Default: FALSE}

\item{sleepTime}{PARAM_DESCRIPTION, Default: 1}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\seealso{
\code{\link[purrr]{keep}}

Other query functions: 
\code{\link{executeAthena}()},
\code{\link{queryAncestors}()},
\code{\link{queryCode}()},
\code{\link{queryConceptChildren}()},
\code{\link{queryConceptClassRelationships}()},
\code{\link{queryConceptId}()},
\code{\link{queryDescendants}()},
\code{\link{queryPhraseExactSynonym}()},
\code{\link{queryPhraseExact}()},
\code{\link{queryPhraseLikeSynonym}()},
\code{\link{queryPhraseLike}()},
\code{\link{queryPhraseStringSynonym}()},
\code{\link{queryPhraseString}()},
\code{\link{queryPhrase}()},
\code{\link{queryRelationships}()},
\code{\link{querySynonyms}()},
\code{\link{queryVocabularyRelationships}()}
}
\concept{query functions}
