% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_functions.R
\name{queryPhraseExact}
\alias{queryPhraseExact}
\title{Query Exact Phrase}
\usage{
queryPhraseExact(
  schema,
  phrase,
  caseInsensitive,
  conn = NULL,
  cache_only = FALSE,
  skip_cache = FALSE,
  override_cache = FALSE,
  render_sql = FALSE,
  verbose = FALSE,
  sleepTime = 1
)
}
\description{
Query Exact Phrase
}
