% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_functions.R
\name{queryPhraseExactSynonym}
\alias{queryPhraseExactSynonym}
\title{Query Concept Synonym Table}
\usage{
queryPhraseExactSynonym(
  schema,
  case_insensitive = TRUE,
  phrase,
  conn = NULL,
  cache_only = FALSE,
  skip_cache = FALSE,
  override_cache = FALSE,
  render_sql = FALSE,
  verbose = FALSE,
  sleepTime = 1
)
}
\description{
Query Concept Synonym Table
}
