% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_functions.R
\name{queryPhraseLike}
\alias{queryPhraseLike}
\title{Query Like Phrase}
\usage{
queryPhraseLike(
  schema,
  phrase,
  caseInsensitive,
  conn = NULL,
  render_sql = TRUE,
  cache_resultset = TRUE,
  override_cache = FALSE,
  verbose = FALSE
)
}
\description{
Query Like Phrase
}
