% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_functions.R
\name{queryPhraseLikeSynonym}
\alias{queryPhraseLikeSynonym}
\title{Query Concept Synonym Table for Concepts Like Phrase}
\usage{
queryPhraseLikeSynonym(
  schema,
  caseInsensitive = TRUE,
  phrase,
  conn = NULL,
  cache_only = FALSE,
  skip_cache = FALSE,
  override_cache = FALSE,
  render_sql = FALSE,
  verbose = FALSE,
  sleepTime = 1
)
}
\description{
Query Concept Synonym Table for Concepts Like Phrase
}
