% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_functions.R
\name{queryPhraseStringSynonym}
\alias{queryPhraseStringSynonym}
\title{Query Concept Synonym Table for a Phrase Split into Strings}
\usage{
queryPhraseStringSynonym(
  schema,
  caseInsensitive = TRUE,
  phrase,
  split = " ",
  conn = NULL,
  cache_only = FALSE,
  skip_cache = FALSE,
  override_cache = FALSE,
  render_sql = FALSE,
  verbose = FALSE,
  sleepTime = 1
)
}
\description{
Query Concept Synonym Table for a Phrase Split into Strings
}
