% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_functions.R
\name{queryVocabularyRelationships}
\alias{queryVocabularyRelationships}
\title{Vocabulary Relationships}
\usage{
queryVocabularyRelationships(
  vocabulary_id,
  verbose = FALSE,
  conn = NULL,
  cache_only = FALSE,
  skip_cache = FALSE,
  override_cache = FALSE,
  render_sql = FALSE,
  sleepTime = 1
)
}
\value{
dataframe derived from the concept_relationship table where concept_id_1 are all the concepts belonging to the vocabulary_id argument.
}
\description{
This function retrieves all the relationships of a given vocabulary, identified by the OMOP vocabulary_id.
}
