% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_functions.R
\name{rank_word_frequency}
\alias{rank_word_frequency}
\title{FUNCTION_TITLE}
\usage{
rank_word_frequency(data, concept_col, words_col)
}
\arguments{
\item{data}{PARAM_DESCRIPTION}

\item{concept_col}{PARAM_DESCRIPTION}

\item{words_col}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
\seealso{
\code{\link[dplyr]{tidyeval-compat}},\code{\link[dplyr]{count}},\code{\link[dplyr]{mutate-joins}},\code{\link[dplyr]{group_by_all}},\code{\link[dplyr]{arrange_all}},\code{\link[dplyr]{mutate}},\code{\link[dplyr]{group_by}},\code{\link[dplyr]{select}}
}
