% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{read_cdm_wiki_table}
\alias{read_cdm_wiki_table}
\title{Read GitHub Wiki}
\usage{
read_cdm_wiki_table(expiration_days = 180)
}
\arguments{
\item{expiration_days}{If the cached file was created longer than this time
period in days, the GitHub wiki is re-read and cached. Default: 180}
}
\description{
Read the tables from OHDI's GitHub Wiki at
\url{https://ohdsi.github.io/CommonDataModel/cdm60.html} to annotate
resultsets.
}
\seealso{
\code{\link[R.cache]{findCache}}
 \code{\link[xml2]{read_xml}}
 \code{\link[rvest]{html_nodes}},\code{\link[rvest]{html_table}},\code{\link[rvest]{html_text}}
}
