% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relationship_triplet.R
\name{relationship_triplet}
\alias{relationship_triplet}
\title{Relationship triplet}
\usage{
relationship_triplet(
  .data,
  .col = NULL,
  include_count = TRUE,
  omop = FALSE,
  omop_schema = "omop_vocabulary"
)
}
\description{
Relationship triplet
}
