% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renameAthenaDB.R
\name{renameAthenaDB}
\alias{renameAthenaDB}
\title{Rename the current Athena DB}
\usage{
renameAthenaDB(local_dbname)
}
\arguments{
\item{local_dbname}{Name of a local database other than Athena to connect to to execute the rename SQL statements.}
}
\description{
Rename the current Athena DB
}
