% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renderConceptClassRelationships.R
\name{renderConceptClassRelationships}
\alias{renderConceptClassRelationships}
\title{Query Concept Class Relationships}
\usage{
renderConceptClassRelationships(
  vocabulary_id_1,
  vocabulary_id_2 = NULL,
  schema = NULL
)
}
\arguments{
\item{vocabulary_id_1}{single vocabulary_id for the first vocabulary (Subject)}

\item{vocabulary_id_2}{single vocabulary_id for the second vocabulary (Object). If vocabulary_id_2 is NULL, vocabulary_id_2 is set to vocabulary_id_1.}
}
\value{
A dataframe with 3 columns: 1) concept_class_id of vocabulary_1 with name as "{vocabulary_id_1}_concept_class_id" unless vocabulary_id_2 is NULL, in which case it will be concept_class_id_1 2) relationship_id from the CONCEPT_RELATIONSHIP table, and 3) concept_class_id of vocabulary_2 with name as "{vocabulary_id_2}_concept_class_id" unless vocabulary_id_2 is NULL, in which case it will be concept_class_id_2.
}
\description{
This function retrieves the Subject-Predicate-Object triplet from the CONCEPT_RELATIONSHIP table between 2 vocabularies.
}
