% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprec-render.R
\name{renderQueryAncestors}
\alias{renderQueryAncestors}
\title{Render SQL to Query for Ancestors}
\usage{
renderQueryAncestors(
  descendant_concept_ids,
  vocab_schema,
  min_levels_of_separation = NULL,
  max_levels_of_separation = NULL
)
}
\description{
Render SQL to Query for Ancestors
}
