% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprec-return_valueset_functions.R
\name{return_omop_vocabs}
\alias{return_omop_vocabs}
\title{Returns all possible OMOP domains in Athena}
\usage{
return_omop_vocabs()
}
\value{
dataframe of unique `domain_id` variables from the `concept` table
}
\description{
Returns all possible OMOP domains in Athena
}
