% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapsibleTrees_functions.R
\name{return_relationship_ids}
\alias{return_relationship_ids}
\title{Return all Hierarchical Relationship Ids}
\usage{
return_relationship_ids(vocabSchema, conn = NULL)
}
\description{
Return all Hierarchical Relationship Ids
}
