% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_out_uncommon_loinc_systems.R
\name{return_uncommon_loinc_systems}
\alias{return_uncommon_loinc_systems}
\title{Filter out uncommon LOINC Systems}
\usage{
return_uncommon_loinc_systems(dataframe, loinc_systems_col)
}
\description{
Filter out uncommon LOINC Systems
}
