% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterValid.R
\name{return_valid}
\alias{return_valid}
\title{Returns all possible OMOP domains in Athena}
\usage{
return_valid(.data, rm_columns = TRUE)
}
\value{
.data of unique `domain_id` variables from the `concept` table. If `rm_columns` = TRUE, removes all the columns that are irrelevant if all valids are filtered for (date, invalid_reason column itself.)
}
\description{
Returns all possible OMOP domains in Athena
}
