% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_cdm_for_concept_id.R
\name{search_cdm_for_concept_id}
\alias{search_cdm_for_concept_id}
\title{Search CDM instance for a concept_id}
\usage{
search_cdm_for_concept_id(conn, concept_id, schema = NULL)
}
\description{
Search CDM instance for a concept_id
}
