% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_chemidplus_rn.R
\name{search_chemidplus_rn}
\alias{search_chemidplus_rn}
\title{Find a ChemiDPlus Registry Number}
\usage{
search_chemidplus_rn(phrase, type = "contains")
}
\description{
Find the Registry Number/s (rn) for a drug
}
