% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x_search_phrase.R
\name{search_ends_with_phrase}
\alias{search_ends_with_phrase}
\title{Search Concepts that End With a Phrase}
\usage{
search_ends_with_phrase(
  phrase,
  case_insensitive = TRUE,
  vocab_schema = "omop_vocabulary",
  conn,
  conn_fun = "connectAthena()",
  cache_only = FALSE,
  skip_cache = FALSE,
  override_cache = FALSE,
  cache_resultset = TRUE,
  render_sql = TRUE,
  render_only = FALSE,
  verbose = TRUE,
  sleepTime = 1
)
}
\arguments{
\item{phrase}{String to search.}

\item{case_insensitive}{Should the search ignore case?, Default: TRUE}

\item{vocab_schema}{OMOP Vocabulary schema, Default: 'omop_vocabulary'}

\item{conn}{Connection object. If provided, diverts queries
to the connection instead of the local Athena instance without caching
features.}

\item{cache_only}{Loads from the cache and does not query the
database. A NULL object is returned if a resultset was not cached.}

\item{skip_cache}{Skip the caching altogether and directly
query the database.}

\item{override_cache}{If TRUE, the cache will not be loaded and will
be overwritten by a new query. For override_cache to take effect,
skip_cache should be FALSE.}

\item{cache_resultset}{(deprecated) If TRUE, the resultset from the
query will first be loaded from the cache. The query will be executed if a
cached resultset is not retrieved for this particular query, after which the
resultset will be cached. If FALSE, Athena or conn will be directly queried
without any caching operations.}

\item{render_sql}{If TRUE, the SQL will be printed back in the
console prior to execution. Default: FALSE}

\item{verbose}{If TRUE, prints loading and querying operations
messages to the console. Default: FALSE}

\item{sleepTime}{Argument for `Sys.sleep()` in between queries to
allow for halting function execution, especially in cases where other chariot
functions are executing multiple queries in succession and require
cancellation.}
}
\description{
Search for concepts that end with a phrase in the Concept Synonym table.
}
\examples{
library(tidyverse)
library(chariot)

# Search the OMOP Vocabulary for an exact match to the phrase
# "Myocardial Infarction" insensitive to case.
search_exact_phrase(phrase = "Myocardial Infarction",
                    case_insensitive = TRUE)


# Repeat with case sensitivity
search_exact_phrase(phrase = "Myocardial Infarction",
                    case_insensitive = FALSE)


# Search concepts that contain the phrase "Myocardial Infarction"
search_like_phrase(phrase = "Myocardial infarction")
search_like_phrase(phrase = "Myocardial infarction",
                   case_insensitive = FALSE)

# Search concepts that start with or end with the phrase "Myocardial Infarction"
search_starts_with_phrase(phrase = "Myocardial infarction")
search_ends_with_phrase(phrase = "Myocardial infarction")

# Search a phrase by splitting into smaller words that are each included in a
# 'LIKE' clause
search_split_phrase(phrase = "Myocardial infarction",
                    split = " ")
}
\seealso{
\code{\link[SqlRender]{render}}
}
