% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{searchPhraseString}
\alias{searchPhraseString}
\title{FUNCTION_TITLE}
\usage{
searchPhraseString(
  schema,
  caseInsensitive = TRUE,
  phrase,
  split = " ",
  conn = NULL,
  cache_only = FALSE,
  skip_cache = FALSE,
  override_cache = FALSE,
  render_sql = FALSE,
  verbose = FALSE,
  sleepTime = 1
)
}
\arguments{
\item{schema}{PARAM_DESCRIPTION}

\item{caseInsensitive}{PARAM_DESCRIPTION, Default: TRUE}

\item{phrase}{PARAM_DESCRIPTION}

\item{split}{PARAM_DESCRIPTION, Default: ' '}

\item{conn}{PARAM_DESCRIPTION, Default: NULL}

\item{cache_only}{PARAM_DESCRIPTION, Default: FALSE}

\item{skip_cache}{PARAM_DESCRIPTION, Default: FALSE}

\item{override_cache}{PARAM_DESCRIPTION, Default: FALSE}

\item{render_sql}{PARAM_DESCRIPTION, Default: FALSE}

\item{verbose}{PARAM_DESCRIPTION, Default: FALSE}

\item{sleepTime}{PARAM_DESCRIPTION, Default: 1}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\seealso{
\code{\link[pg13]{buildQueryString}}

Other query functions: 
\code{\link{executeAthena}()},
\code{\link{queryAncestors}()},
\code{\link{queryCode}()},
\code{\link{queryConceptClassRelationships}()},
\code{\link{queryConceptId}()},
\code{\link{queryDescendants}()},
\code{\link{queryRelationships}()},
\code{\link{querySynonyms}()},
\code{\link{queryVocabularyRelationships}()},
\code{\link{searchPhraseExactSynonym}()},
\code{\link{searchPhraseExact}()},
\code{\link{searchPhraseLikeSynonym}()},
\code{\link{searchPhraseLike}()},
\code{\link{searchPhraseStringSynonym}()},
\code{\link{searchPhrase}()}
}
\concept{query functions}
