% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_functions.R
\name{sendAthena}
\alias{sendAthena}
\title{FUNCTION_TITLE}
\usage{
sendAthena(conn = NULL, sql_statement)
}
\arguments{
\item{conn}{PARAM_DESCRIPTION, Default: NULL}

\item{sql_statement}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[pg13]{send}}
}
