% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_functions.R
\name{separateConceptStrip}
\alias{separateConceptStrip}
\title{Separate Concept Strips by Row}
\usage{
separateConceptStrip(data, ...)
}
\arguments{
\item{data}{A data frame.}

\item{...}{Columns to separate across multiple rows that are passed to \code{\link[tidyr]{separate_rows}}.}
}
\value{
A longer tibble if the merged Concept Strip Column/s have greater than 1 Concept Strips.
}
\description{
This function separates a merged Concept Strip Column into rows by new line \"\\n\".
}
\seealso{
\code{\link[tidyr]{separate_rows}}
}
