% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprec-set_parameters_functions.R
\name{set_search_parameters}
\alias{set_search_parameters}
\title{Create a Search Settings Object}
\usage{
set_search_parameters(obj = "SEARCH_SETTINGS", rm_source_obj = TRUE)
}
\value{
OUTPUT_DESCRIPTION
}
\description{
This function takes the Concept Table fields present in the global environment and aggregates them into a list object that can then be run against a data frame as a whole using the \code{\link{filterSettings}} function.
}
\details{
DETAILS
}
\seealso{
\code{\link[rubix]{map_names_set}}
 \code{\link[purrr]{keep}},\code{\link[purrr]{map}}
}
