% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{stripToLabel}
\alias{stripToLabel}
\title{Convert a Merge Strip to a Label}
\usage{
stripToLabel(data, merge_col, into, remove = FALSE)
}
\description{
Convert a Merge Strip to a Label
}
\keyword{internal}
