% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search-ff.R
\name{tokenized_phrase_search_ff}
\alias{tokenized_phrase_search_ff}
\title{FUNCTION_TITLE}
\usage{
tokenized_phrase_search_ff(
  vocabulary_id,
  domain_id,
  concept_class_id,
  standard_concept,
  invalid_reason = "NULL",
  lowered_match = TRUE
)
}
\arguments{
\item{vocabulary_id}{PARAM_DESCRIPTION}

\item{domain_id}{PARAM_DESCRIPTION}

\item{concept_class_id}{PARAM_DESCRIPTION}

\item{standard_concept}{PARAM_DESCRIPTION}

\item{invalid_reason}{PARAM_DESCRIPTION, Default: 'NULL'}

\item{lowered_match}{PARAM_DESCRIPTION, Default: TRUE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[SqlRender]{render}}
}
