% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprec-format_functions.R
\name{unboxLabel}
\alias{unboxLabel}
\title{Unbox Strip}
\usage{
unboxLabel(data, label_col, row_sep = "\\n", remove = FALSE)
}
\description{
First separates rows by the `row_sep` argument, followed by unmerging the strip
}
\seealso{
\code{\link[tidyr]{separate_rows}}
}
