% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{unboxStrip}
\alias{unboxStrip}
\title{Unbox Strip}
\usage{
unboxStrip(data, strip_col, row_sep = "\\n", remove = FALSE)
}
\description{
First separates rows by the `row_sep` argument, followed by unmerging the strip
}
\seealso{
\code{\link[tidyr]{separate_rows}}
}
