% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unmerge_concepts.R
\name{unmerge_concepts}
\alias{unmerge_concepts}
\title{Unmerge OMOP concept element string}
\usage{
unmerge_concepts(dataframe, concept_col)
}
\arguments{
\item{dataframe}{dataframe}

\item{concept_col}{column that contains the output from the merge_omop_concept_elements function}
}
\description{
This function unmerges an OMOP concept string created by the merge_omop_concept_elements function
}
