% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unmerge_omop_concept_elements.R
\name{unmerge_omop_concept_elements}
\alias{unmerge_omop_concept_elements}
\title{Unmerge OMOP concept element string}
\usage{
unmerge_omop_concept_elements(dataframe, omop_string_col)
}
\arguments{
\item{dataframe}{dataframe}

\item{omop_string_col}{column that contains the output from the merge_omop_concept_elements function}
}
\description{
This function unmerges an OMOP concept string created by the merge_omop_concept_elements function
}
