% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{unmerge_strip}
\alias{unmerge_strip}
\title{Unmerge a Strip}
\usage{
unmerge_strip(
  data,
  strip_col,
  suffix = NULL,
  prefix = NULL,
  remove = TRUE,
  r_trimws = TRUE
)
}
\arguments{
\item{data}{dataframe}

\item{strip_col}{column that contains the merged concept strip}

\item{remove}{remove argument passed to the tidyr extract function. If TRUE, removes strip_col in output.}

\item{r_trimws}{Due to some of the carriage returns in aggregate transformations and other edits in Excel, r_trimws is an argument that if TRUE, trims right whitespace of the freshly unmerged columns for any trailing carriage returns.}
}
\value{
a tibble with all blanks, "NA", and <NA> normalized to NA, with unmerged columns with or without the provided prefix and suffix pasted in postions 1 to 8, followed by the strip column if the remove parameter is FALSE, and the remaining fields present in the input.
}
\description{
This function unmerges an OMOP concept strip created by a 'merge' function using the tidyr extract function. If the input is not a tibble, it will be converted into one with the blanks and "NA" values normalized to `<NA>`. A warning is returned in the console if some concepts fail to unmerge into their respective concept table fields, as determined by all the new column fields having a value of `<NA>` with a non-`<NA>` value in the strip_col instance inputed. Errors will be thrown if the data input already contains columns that will collide with the new columns, the names of which defaults to the names of the original concept table fields: concept_id, concept_name, domain_id, vocabulary_id, concept_class_id, standard_concept, concept_code, invalid_reason. Note that the original concept table fields `valid_start_date` and `valid_end_date` are the only concept table fields are not a requirement in the merge and unmerging process.
}
\seealso{
\code{\link[dplyr]{tidyeval-compat}},\code{\link[dplyr]{mutate_all}},\code{\link[dplyr]{vars}},\code{\link[dplyr]{reexports}},\code{\link[dplyr]{select}},\code{\link[dplyr]{filter_all}},\code{\link[dplyr]{all_vars}}
 \code{\link[tidyr]{extract}}
 \code{\link[tibble]{as_tibble}}
 \code{\link[rubix]{normalize_all_to_na}}
 \code{\link[stringr]{str_remove}},\code{\link[stringr]{str_replace}}
 \code{\link[base]{trimws}}

Other concept format functions: 
\code{\link{filter_at_all_strip}()},
\code{\link{filter_at_any_strip}()},
\code{\link{filter_strip}()},
\code{\link{get_strip}()},
\code{\link{label_to_strip}()},
\code{\link{merge_label}()},
\code{\link{merge_strip}()},
\code{\link{unbox_label}()},
\code{\link{unbox_strip}()},
\code{\link{unmerge_label}()}
}
\concept{concept format functions}
