% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viaMetathesaurusCode.R
\name{viaMetathesaurusCode}
\alias{viaMetathesaurusCode}
\title{Crosswalk a Search by the Metathesaurus CODE}
\usage{
viaMetathesaurusCode(
  phrase,
  caseInsensitive = TRUE,
  conn = NULL,
  athena_schema,
  type = c("like", "exact"),
  verbose = FALSE,
  cache_resultset = TRUE,
  override_cache = FALSE,
  render_sql = TRUE,
  umls_conn = NULL,
  umls_username = Sys.getenv("umls_username"),
  umls_password = Sys.getenv("umls_password")
)
}
\description{
Crosswalk a Search by the Metathesaurus CODE
}
\seealso{
\code{\link[SqlRender]{translate}}
 \code{\link[pg13]{buildQuery}},\code{\link[pg13]{query}}
 \code{\link[stringr]{str_replace}}
 \code{\link[preQL]{connectMySQL5.5}},\code{\link[preQL]{dcMySQL5.5}}
 \code{\link[tibble]{tibble}}
 \code{\link[dplyr]{mutate-joins}},\code{\link[dplyr]{filter}}
 \code{\link[rubix]{filter_for}}
}
