% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x_vocabulary_relationships.R
\name{vocab_lookup_intrarelation}
\alias{vocab_lookup_intrarelation}
\title{Query a Source Vocabulary's Relationships}
\usage{
vocab_lookup_intrarelation(
  vocabulary_id,
  conn,
  conn_fun = "connectAthena()",
  vocab_schema = "omop_vocabulary",
  cache_only = FALSE,
  skip_cache = FALSE,
  override_cache = FALSE,
  cache_resultset = TRUE,
  render_sql = FALSE,
  verbose = FALSE,
  sleepTime = 1
)
}
\description{
Query a Source Vocabulary's Relationships
}
\details{
Query for the Source Vocabulary's non-null relationships in the Concept Relationship Table and all of the OMOP Source Vocabulary and Concept Classes that the relationships are to. For a resultset that filters for relationships to other vocabularies, see \code{\link{query_intervocabulary_relationship}}.
}
\examples{
library(chariot)
library(tidyverse)

# Lookup relationships on the Vocabulary-Concept Class Id axis
vocab_lookup_relationships(vocabulary_id = "HemOnc")

# Relationships between HemOnc concepts only can also be looked up
vocab_lookup_intrarelation(vocabulary_id = "HemOnc")

# Relationships between HemOnc and non-Hemonc Concepts
vocab_lookup_interrelation(vocabulary_id = "HemOnc")
}
\seealso{
\code{\link[SqlRender]{render}}
}
