% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vocabulary_relationships.R
\name{vocab_lookup_intrarelationship}
\alias{vocab_lookup_intrarelationship}
\title{Query a Source Vocabulary's Relationships}
\usage{
vocab_lookup_intrarelationship(
  vocabulary_id,
  conn,
  conn_fun = "connectAthena()",
  vocab_schema = "omop_vocabulary",
  cache_only = FALSE,
  skip_cache = FALSE,
  override_cache = FALSE,
  cache_resultset = TRUE,
  render_sql = FALSE,
  verbose = FALSE,
  sleepTime = 1
)
}
\description{
This family of functions explores the OMOP Vocabulary at the Source Vocabulary level. Lookup functions find details about the given vocabulary while the Query functions query for the inter- and intra-relationships between vocabularies. The concept_class_id for the vocabulary is retained to provide a layer of granularity in the resultsets.
}
\details{
Query for the Source Vocabulary's non-null relationships in the Concept Relationship Table and all of the OMOP Source Vocabulary and Concept Classes that the relationships are to. For a resultset that filters for relationships to other vocabularies, see \code{\link{query_intervocabulary_relationship}}.
}
\seealso{
\code{\link[SqlRender]{render}}
}
