% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x_vocabulary_relationships.R
\name{vocab_lookup_relationships}
\alias{vocab_lookup_relationships}
\title{Lookup a Source Vocabulary's Relationships}
\usage{
vocab_lookup_relationships(
  vocabulary_id,
  conn,
  conn_fun = "connectAthena()",
  vocab_schema = "omop_vocabulary",
  cache_only = FALSE,
  skip_cache = FALSE,
  override_cache = FALSE,
  render_sql = FALSE,
  render_only = FALSE,
  verbose = FALSE,
  sleepTime = 1
)
}
\arguments{
\item{vocabulary_id}{Vector of 1 or more `vocabulary_id` from the OMOP Vocabulary}

\item{conn}{Connection object. If provided, diverts queries
to the connection instead of the local Athena instance without caching
features.}

\item{cache_only}{Loads from the cache and does not query the
database. A NULL object is returned if a resultset was not cached.}

\item{skip_cache}{Skip the caching altogether and directly
query the database.}

\item{override_cache}{If TRUE, the cache will not be loaded and will
be overwritten by a new query. For override_cache to take effect,
skip_cache should be FALSE.}

\item{render_sql}{If TRUE, the SQL will be printed back in the
console prior to execution. Default: FALSE}

\item{verbose}{If TRUE, prints loading and querying operations
messages to the console. Default: FALSE}

\item{sleepTime}{Argument for `Sys.sleep()` in between queries to
allow for halting function execution, especially in cases where other chariot
functions are executing multiple queries in succession and require
cancellation.}
}
\description{
Lookup a Source Vocabulary's Relationships
}
\details{
Lookup all of a Source Vocabulary's non-null relationships in the Concept Relationship Table.
}
\examples{
library(chariot)
library(tidyverse)

# Lookup relationships on the Vocabulary-Concept Class Id axis
vocab_lookup_relationships(vocabulary_id = "HemOnc")

# Relationships between HemOnc concepts only can also be looked up
vocab_lookup_intrarelation(vocabulary_id = "HemOnc")

# Relationships between HemOnc and non-Hemonc Concepts
vocab_lookup_interrelation(vocabulary_id = "HemOnc")
}
\seealso{
\code{\link[SqlRender]{render}}
}
