% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vocabulary_level_functions.R
\name{vocabulary_level_functions}
\alias{vocabulary_level_functions}
\title{Vocabulary Level Queries}
\description{
This family of functions explores the OMOP Vocabulary at the Source Vocabulary level. Lookup functions find details about the given vocabulary while the Query functions query for the inter- and intra-relationships between vocabularies. The concept_class_id for the vocabulary is retained to provide a layer of granularity in the resultsets.
}
\keyword{internal}
