% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_sql_all_vocab_relationships.R
\name{write_sql_all_vocab_relationships}
\alias{write_sql_all_vocab_relationships}
\title{SQL to get all vocabulary relationships}
\usage{
write_sql_all_vocab_relationships(vocabulary_id)
}
\description{
SQL to get all vocabulary relationships
}
