% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_sql_for_ancestors.R
\name{write_sql_for_ancestors}
\alias{write_sql_for_ancestors}
\title{Write SQL to get OMOP Concept Ancestors}
\usage{
write_sql_for_ancestors(
  descendant_concept_ids,
  max_levels_of_separation = NULL
)
}
\arguments{
\item{descendant_concept_ids}{OMOP concept id of the ancestor concept}

\item{max_levels_of_separation}{Integer of the maximum levels of separation}
}
\description{
This function returns the SQL statement that can be used to query on either a cloud or local instance of the OMOP vocabulary.
}
