% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_sql_for_descendants.R
\name{write_sql_for_descendants}
\alias{write_sql_for_descendants}
\title{Write SQL to get OMOP Concept Descendants}
\usage{
write_sql_for_descendants(
  ancestor_concept_ids,
  max_levels_of_separation = NULL
)
}
\arguments{
\item{ancestor_concept_ids}{A vector of 1 or more OMOP concept id of the ancestor concept}

\item{max_levels_of_separation}{Optional. Integer of the maximum levels of separation.}
}
\description{
This function returns the SQL statement that can be used to query on either a cloud or local instance of the OMOP vocabulary.
}
