% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_sql_for_maps_to.R
\name{write_sql_for_maps_to}
\alias{write_sql_for_maps_to}
\title{Search for the "Maps to" relationship of a concept id}
\usage{
write_sql_for_maps_to(concept_id)
}
\arguments{
\item{concept_id}{omop concept_id}
}
\description{
Search for the "Maps to" relationship of a concept id
}
