% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_sql_for_maps_to_relationships.R
\name{write_sql_for_maps_to_relationships}
\alias{write_sql_for_maps_to_relationships}
\title{Search the relationships of a concept id}
\usage{
write_sql_for_maps_to_relationships(concept_id)
}
\arguments{
\item{concept_id}{omop concept_id}
}
\description{
Search the relationships of a concept id
}
