% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_sql_to_search_relationship.R
\name{write_sql_to_get_id_relationship}
\alias{write_sql_to_get_id_relationship}
\title{Search the relationships of a concept id}
\usage{
write_sql_to_get_id_relationship(concept_id)
}
\arguments{
\item{concept_id}{omop concept_id}
}
\description{
Search the relationships of a concept id
}
