% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_sql_to_get_relationships.R
\name{write_sql_to_get_relationships}
\alias{write_sql_to_get_relationships}
\title{Search the relationships of a concept id}
\usage{
write_sql_to_get_relationships(concept_ids)
}
\arguments{
\item{concept_ids}{omop concept_id. For this function, it can also be a vector..}
}
\description{
Search the relationships of a concept id
}
