% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_sql_to_get_synonyms.R
\name{write_sql_to_get_synonyms}
\alias{write_sql_to_get_synonyms}
\title{Get all synonyms of a concept id}
\usage{
write_sql_to_get_synonyms(concept_id)
}
\arguments{
\item{concept_id}{omop concept_id}
}
\description{
Get all synonyms of a concept id
}
