% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapsibleTrees_functions.R
\name{writeConceptRelationshipMapTable}
\alias{writeConceptRelationshipMapTable}
\title{Write a CONCEPT_PARENT Table}
\usage{
writeConceptRelationshipMapTable(vocabSchema, render_sql = TRUE, conn = NULL)
}
\description{
This function creates a subset of the CONCEPT_RELATIONSHIP Table where all forward "Subsumes" relationships are subset. The Concept Ancestor Table is related to the Concept Relationship Table by whether or not the relationship_id in the Concept Relationship Table defines ancestry according to the Relationship Table.
}
