% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concept_id_exists.R
\name{concept_id_exists}
\alias{concept_id_exists}
\title{QA to make sure the concept_id exists in OMOP}
\usage{
concept_id_exists(concept_id)
}
\description{
QA to make sure the concept_id exists in OMOP
}
