% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conn.R
\name{connectAthena}
\alias{connectAthena}
\title{Connect Local Athena}
\usage{
connectAthena(
  user = NULL,
  password = NULL,
  port = 5432,
  server = "localhost/athena",
  verbose = TRUE
)
}
\description{
Connect Local Athena
}
