% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse_athena.R
\name{convenience_functions}
\alias{convenience_functions}
\title{Convenience Functions}
\description{
Convenience functions that simply tasks by allowing them to be executed from the R console such as opening the OMOP Concept page.
}
\keyword{internal}
