% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSettings.R
\name{createSettings}
\alias{createSettings}
\title{Create Settings}
\usage{
createSettings(
  concept_code = NULL,
  vocabulary_id = NULL,
  domain_id = NULL,
  concept_class_id = NULL,
  standard_concept = NULL,
  invalid_reason = NULL
)
}
\description{
Create Settings
}
