% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_lookup_ucum.R
\name{ds_lookup_ucum}
\alias{ds_lookup_ucum}
\title{Lookup the Units for Drug Strength Processing}
\usage{
ds_lookup_ucum(conn, vocabSchema = "omop_vocabulary")
}
\description{
To normalize the drug exposure calculations in the Drug Strength tables,
all units related to rate, commonly seen in extended release formulations,
or time, are excluded from the final calculation.
}
\seealso{
\code{\link[SqlRender]{render}}
}
