% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_process.R
\name{ds_process_drug_table}
\alias{ds_process_drug_table}
\alias{ds_process_map_table}
\title{Process the Records in Drug Strength}
\usage{
ds_process_drug_table(
  conn,
  conn_fun = "connectAthena()",
  vocab_schema = "omop_vocabulary",
  write_schema = "patelm9",
  verbose = TRUE,
  render_sql = TRUE,
  render_only = FALSE
)

ds_process_map_table(
  conn,
  conn_fun = "connectAthena()",
  vocab_schema = "omop_vocabulary",
  write_schema = "patelm9",
  verbose = TRUE,
  render_sql = TRUE,
  render_only = FALSE
)
}
\value{
A `ds_drug` table in `write_schema`. If it already exists, it will be rewritten.

A `ds_unit_map` table in `write_schema`. If it already exists, it will be
rewritten.
}
\description{
The Drug Strength table is filtered for only valid entries and the 'box_size',
'invalid_reason', 'valid_start_date', and 'valid_end_date' fields are selected
against.

Map `amount_unit_concept_id`, `numerator_unit_concept_id`,
`denominator_unit_concept_id`` their respective concept names and store in
a table.
}
\seealso{
\code{\link[SqlRender]{render}}

\code{\link[SqlRender]{render}}

Other drug strength functions: 
\code{\link{ds_lookup_ucum}()},
\code{\link{ds_process}()},
\code{\link{ds_stage_unit_fraction}()},
\code{\link{ds_stage_value_fraction}()},
\code{\link{ds_stage}()}

Other drug strength functions: 
\code{\link{ds_lookup_ucum}()},
\code{\link{ds_process}()},
\code{\link{ds_stage_unit_fraction}()},
\code{\link{ds_stage_value_fraction}()},
\code{\link{ds_stage}()}
}
\concept{drug strength functions}
\concept{drug strength processing functions}
