% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.R
\name{get_concept}
\alias{get_concept}
\title{Lookup a Concept Id}
\usage{
get_concept(
  concept_id,
  vocab_schema,
  conn,
  cache_only = FALSE,
  skip_cache = FALSE,
  override_cache = FALSE,
  render_sql = FALSE,
  verbose = FALSE,
  sleepTime = 1
)
}
\value{
Concept Class object
}
\description{
Lookup a Concept Id
}
