% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ids_to_integer.R
\name{ids_to_integer}
\alias{ids_to_integer}
\title{Mutate all concept_ids to integer}
\usage{
ids_to_integer(.data)
}
\description{
Takes all the fields with "concept_id" in their names and converts to integer
}
