% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{innerJoin}
\alias{innerJoin}
\title{INNER JOIN an OMOP Vocabulary Table}
\usage{
innerJoin(
  data,
  column = NULL,
  write_schema = "patelm9",
  vocab_schema = "omop_vocabulary",
  vocab_table,
  vocab_column,
  where_vocab_col = NULL,
  where_vocab_col_in = NULL,
  render_sql = TRUE,
  conn = NULL
)
}
\description{
This function executes the join() function with joinType == "INNER".
}
