% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join.R
\name{join}
\alias{join}
\title{Join a R dataframe to Athena's concept table}
\usage{
join(
  data,
  joinType,
  column = NULL,
  write_schema = "patelm9",
  vocab_schema = "omop_vocabulary",
  vocab_table,
  vocab_column,
  where_vocab_col = NULL,
  where_vocab_col_in = NULL,
  verbose = TRUE,
  conn,
  conn_fun = "connectAthena()",
  render_sql = TRUE,
  sleepTime = 1
)
}
\arguments{
\item{data}{dataframe to join}

\item{column}{string of the column name to join on. If NULL, the 1st column is used.}

\item{vocab_column}{name of column to join dataframe on. Defaults to concept ID.}
}
\description{
This function has better performance than the WHERE IN statement for larger searches. A table in the format of "v{unpunctuated timestamp}" is written to the local Athena. A join is performed with a concept table. The table is subsequently dropped. If a dataframe column is not provided as a the column to join the concept table on, the 1st column will be used by default.
}
