% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join.R
\name{join}
\alias{join}
\title{Join a dataframe object with Athena's concept table}
\usage{
join(
  .data,
  joinType,
  column = NULL,
  athena_schema,
  athena_table,
  athena_column,
  where_athena_col = NULL,
  where_athena_col_in = NULL,
  verbose = FALSE,
  conn = NULL,
  render_sql = FALSE,
  sleepTime = 1,
  ...
)
}
\arguments{
\item{column}{string of the column name to join on. If NULL, the 1st column is used.}

\item{athena_column}{name of column to join dataframe on. Defaults to concept ID.}

\item{dataframe}{dataframe to join}
}
\description{
This function has better performance than the WHERE IN statement for larger searches. A table in the format of "v{unpunctuated timestamp}" is written to the local Athena. A join is performed with a concept table. The table is subsequently dropped. If a dataframe column is not provided as a the column to join the concept table on, the 1st column will be used by default.
}
