% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprec-format_functions.R
\name{labelToStrip}
\alias{labelToStrip}
\title{Concert a Label Column to a Merge Column}
\usage{
labelToStrip(data, label_col, into, remove = FALSE)
}
\description{
Concert a Label Column to a Merge Column
}
