% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/left_join_for_ancestors.R
\name{left_join_for_ancestors}
\alias{left_join_for_ancestors}
\title{Left Join Ancestors to a dataframe of concept_ids}
\usage{
left_join_for_ancestors(dataframe, descendant_id_column = NULL)
}
\description{
Left Join Ancestors to a dataframe of concept_ids
}
