% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/left_join_for_ancestors.R
\name{left_join_for_ancestors}
\alias{left_join_for_ancestors}
\title{Left Join Ancestors to a dataframe of concept_ids}
\usage{
left_join_for_ancestors(
  .data,
  .descendant_id_column = NULL,
  level = NULL,
  omop = FALSE,
  omop_schema = "omop_vocabulary"
)
}
\description{
Left Join Ancestors to a dataframe of concept_ids
}
