% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join.R
\name{leftJoin}
\alias{leftJoin}
\title{LEFT JOIN an OMOP Vocabulary Table}
\usage{
leftJoin(
  data,
  column = NULL,
  write_schema = "patelm9",
  vocab_schema = "omop_vocabulary",
  vocab_table,
  vocab_column,
  where_vocab_col = NULL,
  where_vocab_col_in = NULL,
  verbose = FALSE,
  conn = NULL,
  render_sql = FALSE,
  sleepTime = 1
)
}
\description{
This function executes the join() function with joinType == "LEFT".
}
