% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprec-join.R
\name{leftJoinFoChildren}
\alias{leftJoinFoChildren}
\title{LEFT JOIN the Concept Parent Table}
\usage{
leftJoinFoChildren(
  data,
  vocab_schema,
  parent_id_column = NULL,
  render_sql = TRUE,
  conn = NULL
)
}
\description{
LEFT JOIN the Concept Parent Table
}
