% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leftJoinForAncestors.R
\name{leftJoinForAncestors}
\alias{leftJoinForAncestors}
\title{Left Join a Dataframe to the Concept Ancestor Table}
\usage{
leftJoinForAncestors(
  .data,
  athena_schema = "public",
  descendant_id_column = NULL,
  whereLevelIn = NULL,
  whereLevelType = NULL,
  override_cache = FALSE,
  print_sql = TRUE
)
}
\description{
Left Join a Dataframe to the Concept Ancestor Table
}
