% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join.R
\name{leftJoinForDescendants}
\alias{leftJoinForDescendants}
\title{Left Join a data frame to the Concept Ancestor Table}
\usage{
leftJoinForDescendants(
  data,
  write_schema = "patelm9",
  vocab_schema = "omop_vocabulary",
  ancestor_id_column = NULL,
  whereLevelIn = NULL,
  whereLevelType = NULL,
  verbose = TRUE,
  conn,
  render_sql = TRUE,
  sleepTime = 1
)
}
\arguments{
\item{data}{PARAM_DESCRIPTION}

\item{vocab_schema}{Default: 'public'}

\item{ancestor_id_column}{Default: NULL}

\item{whereLevelIn}{Default: NULL}

\item{whereLevelType}{Default: NULL}

\item{conn}{Default: NULL}

\item{render_sql}{Default: TRUE}
}
\description{
FUNCTION_DESCRIPTION
}
\seealso{
\code{\link[dplyr]{select}},\code{\link[dplyr]{distinct}},\code{\link[dplyr]{mutate-joins}}
 \code{\link[rubix]{rename_all_with_prefix}}
}
