% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join.R
\name{leftJoinForParents}
\alias{leftJoinForParents}
\title{LEFT JOIN the Concept Parent Table}
\usage{
leftJoinForParents(
  data,
  vocab_schema,
  child_id_column = NULL,
  render_sql = TRUE,
  conn = NULL
)
}
\description{
LEFT JOIN the Concept Parent Table
}
