% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leftJoinRelationship.R
\name{leftJoinRelationship}
\alias{leftJoinRelationship}
\title{Left Join Relationship}
\usage{
leftJoinRelationship(.data, column = NULL, athena_schema = "public")
}
\description{
Left Join Relationship
}
