% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leftJoinSynonymId.R
\name{leftJoinSynonymId}
\alias{leftJoinSynonymId}
\title{Join a data frame with the Concept Synonym Table}
\usage{
leftJoinSynonymId(
  .data,
  column = NULL,
  athena_schema,
  verbose = FALSE,
  conn = NULL,
  render_sql = FALSE,
  sleepTime = 1,
  ...
)
}
\arguments{
\item{.data}{A data frame}

\item{column}{Data frame column that the join will be performed on. If NULL, defaults to the column in position 1 of the data frame.}

\item{athena_schema}{Schema of the OMOP Vocabulary Tables}

\item{verbose}{If TRUE, prints whether the cache is being loaded or being actively queried in the Postgres database, Default: FALSE}

\item{conn}{PARAM_DESCRIPTION, Default: NULL}

\item{render_sql}{If TRUE, will print the SQL to the console before executing. Default: FALSE}

\item{sleepTime}{Argument in seconds passed to the `Sys.sleep()` function at the end of query, Default: 1}

\item{...}{Additional arguments passed to the `queryAthena()` function.}
}
\value{
A data frame
}
\description{
This function has better performance than the WHERE IN statement for larger searches. A table in the format of "v{unpunctuated timestamp}" is written to the local Athena. A join is performed with a concept table. The table is subsequently dropped. If a dataframe column is not provided as a the column to join the concept table on, the 1st column will be used by default.
}
\seealso{
\code{\link[dplyr]{select}}
}
