% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{makeLabel}
\alias{makeLabel}
\title{Make Label Column}
\usage{
makeLabel(data, into, prefix = NULL, suffix = NULL, remove = FALSE)
}
\description{
A Label is in the format of "{concept_id} concept_name". It is less comprehensive than a merged strip using the merge_concepts function, but more human readable when interfacing with others.
}
