% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{merge_label}
\alias{merge_label}
\title{Make Label Column}
\usage{
merge_label(data, into, prefix = NULL, suffix = NULL, remove = TRUE)
}
\description{
A Label is in the format of "{concept_id} concept_name". It is less comprehensive than a merged strip using the merge_concepts function, but more human readable when interfacing with others.
}
\seealso{
\code{\link[tidyr]{unite}}
 \code{\link[dplyr]{reexports}},\code{\link[dplyr]{mutate_all}},\code{\link[dplyr]{vars}}
}
