% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_relationship_id.R
\name{pivot_relationship_id}
\alias{pivot_relationship_id}
\title{Get all Loinc System (Specimen) Types for a Lab}
\usage{
pivot_relationship_id(dataframe, concept_id_col = NULL, include_count = TRUE)
}
\arguments{
\item{dataframe}{input data}

\item{concept_id_col}{The column in dataframe that points to the concept_id. If NULL, defaults to "concept_id".}
}
\description{
This function takes a dataframe and mutates an additional column providing the specimen type based on the "Has system" relationship id.
}
