% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_relative.R
\name{pivot_relative}
\alias{pivot_relative}
\title{Pivot all relatives of a set of concepts}
\usage{
pivot_relative(
  .data,
  id_column = NULL,
  names_from,
  include_count = TRUE,
  omop = FALSE,
  omop_schema = "omop_vocabulary"
)
}
\arguments{
\item{names_from}{concept table column to be pivoted on}

\item{concept_id_column}{The column in dataframe that points to the concept_id. If NULL, defaults to "concept_id".}

\item{dataframe}{input data}
}
\description{
This function takes a dataframe and mutates an additional column providing the specimen type based on the "Has system" relationship id.
}
\examples{
Random immunosuppressant concept ids
immunosuppressant_concept_ids <- c("35807335","35807331", "21603616", "21600651", "21605199", "21602723") 
}
