% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_ancestors.R
\name{query_ancestors}
\alias{query_ancestors}
\title{Query ancestors for a given concept_id}
\usage{
query_ancestors(
  descendant_concept_ids,
  max_levels_of_separation = NULL,
  plot = FALSE
)
}
\description{
Query ancestors for a given concept_id
}
