% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_descendants.R
\name{query_descendants}
\alias{query_descendants}
\title{Query descendants for a given concept_id}
\usage{
query_descendants(ancestor_concept_id, max_levels_of_separation = NULL)
}
\arguments{
\item{ancestor_concept_id}{OMOP concept id of the ancestor concept}

\item{max_levels_of_separation}{Optional. Integer of the maximum levels of separation.}
}
\description{
Query descendants for a given concept_id
}
