% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dep_query_phrases.R, R/query_phrases.R
\name{query_phrases}
\alias{query_phrases}
\title{Query more than 1 search terms}
\usage{
query_phrases(
  phrases,
  limit = NULL,
  type = c("like", "exact"),
  case_insensitive = TRUE,
  set_names_to = NULL
)

query_phrases(
  phrases,
  limit = NULL,
  type = c("like", "exact"),
  case_insensitive = TRUE,
  set_names_to = NULL
)
}
\arguments{
\item{phrases}{vector of phrases to iterate the query_phrase function on}

\item{...}{vector of terms to query}
}
\value{
resultset as a dataframe with all column types as character and trimmed white space
}
\description{
Query more than 1 search terms

Iterate query_phrase function over a vector of phrases
}
