% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_relationships.R
\name{query_relationships}
\alias{query_relationships}
\title{Get all relationships for a concept_id or a vector of concept_ids}
\usage{
query_relationships(concept_ids)
}
\description{
Get all relationships for a concept_id or a vector of concept_ids
}
