% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_relationship.R
\name{query_relationships}
\alias{query_relationships}
\title{Query a Relationship}
\usage{
query_relationships(
  relationship_ids,
  conn,
  conn_fun = "connectAthena()",
  vocab_schema = "omop_vocabulary",
  cache_only = FALSE,
  skip_cache = FALSE,
  override_cache = FALSE,
  render_sql = FALSE,
  verbose = FALSE,
  sleepTime = 1
)
}
\arguments{
\item{...}{Relationship ids.}
}
\description{
Returns a query for a relationship_id.
}
\seealso{
\code{\link[cli]{cli_rule}}
 \code{\link[rlang]{list2}}
 \code{\link[SqlRender]{render}}
}
