% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryAthena.R
\name{queryAthena}
\alias{queryAthena}
\title{Generic query function}
\usage{
queryAthena(
  sql_statement,
  verbose = FALSE,
  cache_resultset = TRUE,
  override_cache = FALSE
)
}
\arguments{
\item{sql_statement}{SQL statement}
}
\description{
This is the low level function that is wrapped for more specific queries. The resultset is cached using the R.cache package
}
