% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute.R, R/query_functions.R
\name{executeAthena}
\alias{executeAthena}
\alias{queryAthena}
\title{Query the Athena Postgres Database}
\usage{
executeAthena(
  sql_statement,
  conn = NULL,
  cache_only = FALSE,
  skip_cache = FALSE,
  override_cache = FALSE,
  cache_resultset = TRUE,
  render_sql = FALSE,
  verbose = FALSE,
  sleepTime = 1
)

queryAthena(
  sql_statement,
  conn = NULL,
  conn_fun = "connectAthena()",
  cache_only = FALSE,
  skip_cache = FALSE,
  override_cache = FALSE,
  cache_resultset = TRUE,
  render_sql = TRUE,
  verbose = TRUE,
  sleepTime = 1
)
}
\arguments{
\item{sql_statement}{SQL query}

\item{conn}{Connection object. If provided, diverts queries to the connection instead of the local Athena instance without caching features.}

\item{cache_only}{Loads from the cache and does not query the database. A NULL object is returned if a resultset was not cached.}

\item{skip_cache}{Skip the caching altogether and directly query the database.}

\item{override_cache}{If TRUE, the cache will not be loaded and will be overwritten by a new query. For override_cache to take effect, skip_cache should be FALSE.}

\item{cache_resultset}{(deprecated) If TRUE, the resultset from the query will first be loaded from the cache. The query will be executed if a cached resultset is not retrieved for this particular query, after which the resultset will be cached. If FALSE, Athena or conn will be directly queried without any caching operations.}

\item{render_sql}{If TRUE, the SQL will be printed back in the console prior to execution. Default: FALSE}

\item{verbose}{If TRUE, prints loading and querying operations messages to the console. Default: FALSE}

\item{sleepTime}{Argument for `Sys.sleep()` in between queries to allow for halting function execution, especially in cases where other chariot functions are executing multiple queries in succession and require cancellation.}
}
\value{
A tibble

A tibble
}
\description{
By default, this function queries a local database named "Athena". If a connection object is passed into the function, the database of the connection object is queried instead. The caching feature is only available when using the built-in connection to Athena.

By default, this function queries a local database named "Athena". If a connection object is passed into the function, the database of the connection object is queried instead. The caching feature is only available when using the built-in connection to Athena.
}
\seealso{
\code{\link[secretary]{typewrite_bold}},\code{\link[secretary]{typewrite}}
 \code{\link[stringr]{str_replace}},\code{\link[stringr]{str_remove}}
 \code{\link[pg13]{query}},\code{\link[pg13]{cacheQuery}},\code{\link[pg13]{loadCachedQuery}}
 \code{\link[tibble]{as_tibble}}

\code{\link[secretary]{typewrite_bold}},\code{\link[secretary]{typewrite}}
 \code{\link[stringr]{str_replace}},\code{\link[stringr]{str_remove}}
 \code{\link[pg13]{query}},\code{\link[pg13]{cacheQuery}},\code{\link[pg13]{loadCachedQuery}}
 \code{\link[tibble]{as_tibble}}
}
