% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-query_functions.R, R/query_functions.R
\name{queryCode}
\alias{queryCode}
\title{FUNCTION_TITLE}
\usage{
queryCode(
  code,
  schema,
  caseInsensitive = TRUE,
  limit = NULL,
  verbose = FALSE,
  conn = NULL,
  cache_only = FALSE,
  skip_cache = FALSE,
  override_cache = FALSE,
  render_sql = FALSE,
  sleepTime = 1
)

queryCode(
  code,
  schema,
  caseInsensitive = TRUE,
  limit = NULL,
  verbose = FALSE,
  conn = NULL,
  cache_only = FALSE,
  skip_cache = FALSE,
  override_cache = FALSE,
  render_sql = FALSE,
  sleepTime = 1
)
}
\arguments{
\item{code}{PARAM_DESCRIPTION}

\item{schema}{PARAM_DESCRIPTION}

\item{caseInsensitive}{PARAM_DESCRIPTION, Default: TRUE}

\item{limit}{PARAM_DESCRIPTION, Default: NULL}

\item{verbose}{PARAM_DESCRIPTION, Default: FALSE}

\item{conn}{PARAM_DESCRIPTION, Default: NULL}

\item{cache_only}{PARAM_DESCRIPTION, Default: FALSE}

\item{skip_cache}{PARAM_DESCRIPTION, Default: FALSE}

\item{override_cache}{PARAM_DESCRIPTION, Default: FALSE}

\item{render_sql}{PARAM_DESCRIPTION, Default: FALSE}

\item{sleepTime}{PARAM_DESCRIPTION, Default: 1}

\item{...}{vector of codes to collectively feed into the LIKE sql statement}
}
\value{
OUTPUT_DESCRIPTION

resultset as a dataframe with all column types as character and trimmed white space
}
\description{
FUNCTION_DESCRIPTION
}
\seealso{
\code{\link[pg13]{buildQuery}}

Other query functions: 
\code{\link{executeAthena}()},
\code{\link{queryAncestors}()},
\code{\link{queryConceptChildren}()},
\code{\link{queryConceptClassRelationships}()},
\code{\link{queryConceptId}()},
\code{\link{queryDescendants}()},
\code{\link{queryFamilyTree}()},
\code{\link{queryPhraseExactSynonym}()},
\code{\link{queryPhraseExact}()},
\code{\link{queryPhraseLikeSynonym}()},
\code{\link{queryPhraseLike}()},
\code{\link{queryPhraseStringSynonym}()},
\code{\link{queryPhraseString}()},
\code{\link{queryPhrase}()},
\code{\link{queryRelationships}()},
\code{\link{querySynonyms}()},
\code{\link{queryVocabularyRelationships}()}
}
\concept{query functions}
