% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_functions.R
\name{queryConceptChildren}
\alias{queryConceptChildren}
\title{Query concept children}
\usage{
queryConceptChildren(
  parent_id,
  generations = 1,
  override_cache = FALSE,
  verbose = FALSE,
  cache_resultset = TRUE,
  conn = NULL,
  render_sql = TRUE,
  schema
)
}
\description{
Query concept children
}
