% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-query_functions.R, R/query_functions.R
\name{queryConceptClassRelationships}
\alias{queryConceptClassRelationships}
\title{FUNCTION_TITLE}
\usage{
queryConceptClassRelationships(
  vocabulary_id_1,
  vocabulary_id_2 = NULL,
  schema = NULL,
  conn = NULL,
  cache_only = FALSE,
  skip_cache = FALSE,
  override_cache = FALSE,
  render_sql = FALSE,
  verbose = FALSE,
  sleepTime = 1
)

queryConceptClassRelationships(
  vocabulary_id_1,
  vocabulary_id_2 = NULL,
  schema = NULL,
  conn = NULL,
  cache_only = FALSE,
  skip_cache = FALSE,
  override_cache = FALSE,
  render_sql = FALSE,
  verbose = FALSE,
  sleepTime = 1
)
}
\arguments{
\item{vocabulary_id_1}{single vocabulary_id for the first vocabulary (Subject)}

\item{vocabulary_id_2}{single vocabulary_id for the second vocabulary (Object). If vocabulary_id_2 is NULL, vocabulary_id_2 is set to vocabulary_id_1.}

\item{schema}{PARAM_DESCRIPTION, Default: NULL}

\item{conn}{PARAM_DESCRIPTION, Default: NULL}

\item{cache_only}{PARAM_DESCRIPTION, Default: FALSE}

\item{skip_cache}{PARAM_DESCRIPTION, Default: FALSE}

\item{override_cache}{PARAM_DESCRIPTION, Default: FALSE}

\item{render_sql}{PARAM_DESCRIPTION, Default: FALSE}

\item{verbose}{PARAM_DESCRIPTION, Default: FALSE}

\item{sleepTime}{PARAM_DESCRIPTION, Default: 1}
}
\value{
OUTPUT_DESCRIPTION

A dataframe with 3 columns: 1) concept_class_id of vocabulary_1 with name as "{vocabulary_id_1}_concept_class_id" unless vocabulary_id_2 is NULL, in which case it will be concept_class_id_1 2) relationship_id from the CONCEPT_RELATIONSHIP table, and 3) concept_class_id of vocabulary_2 with name as "{vocabulary_id_2}_concept_class_id" unless vocabulary_id_2 is NULL, in which case it will be concept_class_id_2.
}
\description{
FUNCTION_DESCRIPTION

This function retrieves the Subject-Predicate-Object triplet from the CONCEPT_RELATIONSHIP table between 2 vocabularies.
}
\seealso{
Other query functions: 
\code{\link{executeAthena}()},
\code{\link{queryAncestors}()},
\code{\link{queryCode}()},
\code{\link{queryConceptChildren}()},
\code{\link{queryConceptId}()},
\code{\link{queryDescendants}()},
\code{\link{queryFamilyTree}()},
\code{\link{queryPhraseExactSynonym}()},
\code{\link{queryPhraseExact}()},
\code{\link{queryPhraseLikeSynonym}()},
\code{\link{queryPhraseLike}()},
\code{\link{queryPhraseStringSynonym}()},
\code{\link{queryPhraseString}()},
\code{\link{queryPhrase}()},
\code{\link{queryRelationships}()},
\code{\link{querySynonyms}()},
\code{\link{queryVocabularyRelationships}()}
}
\concept{query functions}
