% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_functions.R
\name{queryConceptId}
\alias{queryConceptId}
\title{Lookup a concept id in Athena}
\usage{
queryConceptId(
  concept_ids,
  schema,
  conn = NULL,
  cache_only = FALSE,
  skip_cache = FALSE,
  override_cache = FALSE,
  render_sql = FALSE,
  verbose = FALSE,
  sleepTime = 1
)
}
\description{
Lookup a concept id in Athena
}
