% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryConceptParent.R
\name{queryConceptParent}
\alias{queryConceptParent}
\title{Query concept children}
\usage{
queryConceptParent(
  child_id,
  generations = 1,
  override_cache = FALSE,
  verbose = FALSE,
  cache_resultset = TRUE,
  conn = NULL,
  render_sql = TRUE,
  schema
)
}
\description{
Query concept children
}
