% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_functions.R
\name{queryDescendants}
\alias{queryDescendants}
\title{Query descendants for a given concept_id}
\usage{
queryDescendants(
  ancestor_concept_ids,
  schema,
  min_levels_of_separation = NULL,
  max_levels_of_separation = NULL,
  conn = NULL,
  cache_only = FALSE,
  skip_cache = FALSE,
  override_cache = FALSE,
  render_sql = FALSE,
  verbose = FALSE,
  sleepTime = 1
)
}
\description{
Query descendants for a given concept_id
}
