% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryDescendants.R
\name{queryDescendants}
\alias{queryDescendants}
\title{Query descendants for a given concept_id}
\usage{
queryDescendants(
  ancestor_concept_ids,
  schema,
  min_levels_of_separation = NULL,
  max_levels_of_separation = NULL,
  verbose = FALSE,
  cache_resultset = TRUE,
  override_cache = FALSE,
  conn = NULL,
  render_sql = FALSE,
  sleepTime = 1,
  ...
)
}
\description{
Query descendants for a given concept_id
}
