% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_functions.R
\name{queryFamilyTree}
\alias{queryFamilyTree}
\title{Combine Parents and Children Relationships}
\usage{
queryFamilyTree(
  concept_id,
  schema,
  parental_generations = 1,
  child_generations = 1,
  conn = NULL,
  cache_only = FALSE,
  skip_cache = FALSE,
  override_cache = FALSE,
  render_sql = FALSE,
  verbose = FALSE,
  sleepTime = 1
)
}
\description{
Combine Parents and Children Relationships
}
