% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryFamilyTree.R
\name{queryFamilyTree}
\alias{queryFamilyTree}
\title{Combine Parents and Children Relationships}
\usage{
queryFamilyTree(
  concept_id,
  schema,
  parental_generations = 1,
  child_generations = 1,
  verbose = FALSE,
  cache_resultset = TRUE,
  override_cache = FALSE,
  conn = NULL,
  render_sql = FALSE,
  sleepTime = 1,
  ...
)
}
\description{
Combine Parents and Children Relationships
}
