% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_functions.R
\name{queryHemOncRegToAntineo}
\alias{queryHemOncRegToAntineo}
\title{Query a HemOnc Regimen's 'Has antineoplastic' Relationship}
\usage{
queryHemOncRegToAntineo(
  regimen_concept_ids,
  schema = NULL,
  conn = NULL,
  cache_only = FALSE,
  skip_cache = FALSE,
  override_cache = FALSE,
  render_sql = FALSE,
  verbose = FALSE,
  sleepTime = 1
)
}
\description{
Query a HemOnc Regimen's 'Has antineoplastic' Relationship
}
