% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_functions.R
\name{queryPhrase}
\alias{queryPhrase}
\title{Perform a Global Query on a Single Phrase}
\usage{
queryPhrase(
  schema,
  phrase,
  split,
  caseInsensitive = TRUE,
  concept_table_search = c("exact", "string", "like"),
  concept_synonym_table_search = c("exact", "string", "like"),
  conn = NULL,
  cache_only = FALSE,
  skip_cache = FALSE,
  override_cache = FALSE,
  render_sql = FALSE,
  verbose = FALSE,
  sleepTime = 1
)
}
\description{
Perform a Global Query on a Single Phrase
}
