% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_functions.R
\name{queryPhraseExact}
\alias{queryPhraseExact}
\title{Query Exact Phrase}
\usage{
queryPhraseExact(
  schema,
  phrase,
  caseInsensitive,
  conn = NULL,
  render_sql = TRUE,
  cache_resultset = TRUE,
  override_cache = FALSE,
  verbose = FALSE
)
}
\description{
Query Exact Phrase
}
