% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryPhraseExactSynonym.R
\name{queryPhraseExactSynonym}
\alias{queryPhraseExactSynonym}
\title{Query Concept Synonym Table for a Phrase Split into Strings}
\usage{
queryPhraseExactSynonym(
  schema,
  caseInsensitive = TRUE,
  phrase,
  split,
  limit_n = NULL,
  print_sql = TRUE
)
}
\description{
Query Concept Synonym Table for a Phrase Split into Strings
}
