% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryPhraseExactSynonym.R
\name{queryPhraseExactSynonym}
\alias{queryPhraseExactSynonym}
\title{Query Concept Synonym Table}
\usage{
queryPhraseExactSynonym(
  schema,
  caseInsensitive = TRUE,
  phrase,
  limit_n = NULL,
  ...
)
}
\description{
Query Concept Synonym Table
}
