% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-query_functions.R, R/query_functions.R
\name{queryPhraseExactSynonym}
\alias{queryPhraseExactSynonym}
\title{FUNCTION_TITLE}
\usage{
queryPhraseExactSynonym(
  schema,
  case_insensitive = TRUE,
  phrase,
  conn = NULL,
  cache_only = FALSE,
  skip_cache = FALSE,
  override_cache = FALSE,
  render_sql = FALSE,
  verbose = FALSE,
  sleepTime = 1
)

queryPhraseExactSynonym(
  schema,
  case_insensitive = TRUE,
  phrase,
  conn = NULL,
  cache_only = FALSE,
  skip_cache = FALSE,
  override_cache = FALSE,
  render_sql = FALSE,
  verbose = FALSE,
  sleepTime = 1
)
}
\arguments{
\item{schema}{PARAM_DESCRIPTION}

\item{case_insensitive}{PARAM_DESCRIPTION, Default: TRUE}

\item{phrase}{PARAM_DESCRIPTION}

\item{conn}{PARAM_DESCRIPTION, Default: NULL}

\item{cache_only}{PARAM_DESCRIPTION, Default: FALSE}

\item{skip_cache}{PARAM_DESCRIPTION, Default: FALSE}

\item{override_cache}{PARAM_DESCRIPTION, Default: FALSE}

\item{render_sql}{PARAM_DESCRIPTION, Default: FALSE}

\item{verbose}{PARAM_DESCRIPTION, Default: FALSE}

\item{sleepTime}{PARAM_DESCRIPTION, Default: 1}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\seealso{
\code{\link[SqlRender]{render}}

Other query functions: 
\code{\link{executeAthena}()},
\code{\link{queryAncestors}()},
\code{\link{queryCode}()},
\code{\link{queryConceptChildren}()},
\code{\link{queryConceptClassRelationships}()},
\code{\link{queryConceptId}()},
\code{\link{queryDescendants}()},
\code{\link{queryFamilyTree}()},
\code{\link{queryPhraseExact}()},
\code{\link{queryPhraseLikeSynonym}()},
\code{\link{queryPhraseLike}()},
\code{\link{queryPhraseStringSynonym}()},
\code{\link{queryPhraseString}()},
\code{\link{queryPhrase}()},
\code{\link{queryRelationships}()},
\code{\link{querySynonyms}()},
\code{\link{queryVocabularyRelationships}()}
}
\concept{query functions}
