% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_functions.R
\name{queryPhraseExactSynonym}
\alias{queryPhraseExactSynonym}
\title{Query Concept Synonym Table}
\usage{
queryPhraseExactSynonym(
  schema,
  caseInsensitive = TRUE,
  phrase,
  render_sql = TRUE,
  conn = NULL,
  cache_resultset = TRUE,
  override_cache = FALSE,
  verbose = FALSE
)
}
\description{
Query Concept Synonym Table
}
