% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_functions.R
\name{queryPhraseLike}
\alias{queryPhraseLike}
\title{Query Like Phrase}
\usage{
queryPhraseLike(
  schema,
  phrase,
  caseInsensitive,
  conn = NULL,
  cache_only = FALSE,
  skip_cache = FALSE,
  override_cache = FALSE,
  render_sql = FALSE,
  verbose = FALSE,
  sleepTime = 1
)
}
\description{
Query Like Phrase
}
