% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryPhraseLikeSynonym.R
\name{queryPhraseLikeSynonym}
\alias{queryPhraseLikeSynonym}
\title{Query Concept Synonym Table for Concepts Like Phrase}
\usage{
queryPhraseLikeSynonym(
  schema,
  caseInsensitive = TRUE,
  phrase,
  limit_n = NULL,
  print_sql = TRUE
)
}
\description{
Query Concept Synonym Table for Concepts Like Phrase
}
