% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_functions.R
\name{queryPhraseLikeSynonym}
\alias{queryPhraseLikeSynonym}
\title{Query Concept Synonym Table for Concepts Like Phrase}
\usage{
queryPhraseLikeSynonym(
  schema,
  caseInsensitive = TRUE,
  phrase,
  render_sql = TRUE,
  conn = NULL,
  verbose = FALSE,
  cache_resultset = TRUE,
  override_cache = FALSE
)
}
\description{
Query Concept Synonym Table for Concepts Like Phrase
}
