% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryPhraseString.R
\name{queryPhraseString}
\alias{queryPhraseString}
\title{Query Concept Table for a Phrase Split into Strings}
\usage{
queryPhraseString(
  schema,
  caseInsensitive = TRUE,
  phrase,
  split,
  limit_n = NULL,
  print_sql = TRUE
)
}
\description{
Query Concept Table for a Phrase Split into Strings
}
