% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryPhraseString.R
\name{queryPhraseString}
\alias{queryPhraseString}
\title{Query Concept Table for a Phrase Split into Strings}
\usage{
queryPhraseString(
  schema,
  caseInsensitive = TRUE,
  phrase,
  split = " ",
  conn = NULL,
  render_sql = TRUE,
  cache_resultset = TRUE,
  override_cache = FALSE,
  verbose = FALSE
)
}
\description{
Query Concept Table for a Phrase Split into Strings
}
