% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-query_functions.R, R/query_functions.R
\name{queryPhraseStringSynonym}
\alias{queryPhraseStringSynonym}
\title{FUNCTION_TITLE}
\usage{
queryPhraseStringSynonym(
  schema,
  caseInsensitive = TRUE,
  phrase,
  split = " ",
  conn = NULL,
  cache_only = FALSE,
  skip_cache = FALSE,
  override_cache = FALSE,
  render_sql = FALSE,
  verbose = FALSE,
  sleepTime = 1
)

queryPhraseStringSynonym(
  schema,
  caseInsensitive = TRUE,
  phrase,
  split = " ",
  conn = NULL,
  cache_only = FALSE,
  skip_cache = FALSE,
  override_cache = FALSE,
  render_sql = FALSE,
  verbose = FALSE,
  sleepTime = 1
)
}
\arguments{
\item{schema}{PARAM_DESCRIPTION}

\item{caseInsensitive}{PARAM_DESCRIPTION, Default: TRUE}

\item{phrase}{PARAM_DESCRIPTION}

\item{split}{PARAM_DESCRIPTION, Default: ' '}

\item{conn}{PARAM_DESCRIPTION, Default: NULL}

\item{cache_only}{PARAM_DESCRIPTION, Default: FALSE}

\item{skip_cache}{PARAM_DESCRIPTION, Default: FALSE}

\item{override_cache}{PARAM_DESCRIPTION, Default: FALSE}

\item{render_sql}{PARAM_DESCRIPTION, Default: FALSE}

\item{verbose}{PARAM_DESCRIPTION, Default: FALSE}

\item{sleepTime}{PARAM_DESCRIPTION, Default: 1}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\seealso{
\code{\link[pg13]{buildQueryString}}
 \code{\link[dplyr]{rename}},\code{\link[dplyr]{filter}},\code{\link[dplyr]{select}}
 \code{\link[rubix]{group_by_unique_aggregate}}

Other query functions: 
\code{\link{executeAthena}()},
\code{\link{queryAncestors}()},
\code{\link{queryCode}()},
\code{\link{queryConceptChildren}()},
\code{\link{queryConceptClassRelationships}()},
\code{\link{queryConceptId}()},
\code{\link{queryDescendants}()},
\code{\link{queryFamilyTree}()},
\code{\link{queryPhraseExactSynonym}()},
\code{\link{queryPhraseExact}()},
\code{\link{queryPhraseLikeSynonym}()},
\code{\link{queryPhraseLike}()},
\code{\link{queryPhraseString}()},
\code{\link{queryPhrase}()},
\code{\link{queryRelationships}()},
\code{\link{querySynonyms}()},
\code{\link{queryVocabularyRelationships}()}
}
\concept{query functions}
