% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_functions.R
\name{queryPhraseStringSynonym}
\alias{queryPhraseStringSynonym}
\title{Query Concept Synonym Table for a Phrase Split into Strings}
\usage{
queryPhraseStringSynonym(
  schema,
  caseInsensitive = TRUE,
  phrase,
  split,
  render_sql = TRUE,
  verbose = FALSE,
  cache_resultset = TRUE,
  override_cache = FALSE,
  conn = NULL
)
}
\description{
Query Concept Synonym Table for a Phrase Split into Strings
}
