% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/querySynonyms.R
\name{querySynonyms}
\alias{querySynonyms}
\title{Query the synonyms associated with an id}
\usage{
querySynonyms(concept_id, schema = NULL, language_concept_id = 4180186, ...)
}
\arguments{
\item{concept_id}{omop concept_id}

\item{language_concept_id}{language concept id for the synonyms. Defaults to English only.}
}
\description{
Query the synonyms associated with an id
}
