% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_functions.R
\name{querySynonyms}
\alias{querySynonyms}
\title{Query the synonyms associated with an id}
\usage{
querySynonyms(
  concept_id,
  schema = NULL,
  language_concept_id = 4180186,
  conn = NULL,
  cache_only = FALSE,
  skip_cache = FALSE,
  override_cache = FALSE,
  render_sql = FALSE,
  verbose = FALSE,
  sleepTime = 1
)
}
\description{
Query the synonyms associated with an id
}
