% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/querySynonyms.R
\name{querySynonyms}
\alias{querySynonyms}
\title{Query the synonyms associated with an id}
\usage{
querySynonyms(concept_id, schema = NULL, language_concept_id = 4180186, ...)
}
\description{
Query the synonyms associated with an id
}
