% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_functions.R
\name{renderHemOncCompToReg}
\alias{renderHemOncCompToReg}
\title{Render HemOnc Component to Regimen Query}
\usage{
renderHemOncCompToReg(component_concept_ids, schema = NULL)
}
\arguments{
\item{component_concept_ids}{1 or more HemOnc Component Concept Ids.}
}
\value{

}
\description{
This function renders the SQL that retrieves the HemOnc Regimens associated with 1 or more HemOnc Components
}
