% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_functions.R
\name{renderQueryDescendants}
\alias{renderQueryDescendants}
\title{Render SQL to Query for Descendants}
\usage{
renderQueryDescendants(
  ancestor_concept_ids,
  schema,
  min_levels_of_separation = NULL,
  max_levels_of_separation = NULL
)
}
\description{
Render SQL to Query for Descendants
}
