% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renderQueryPhraseExactSynonym.R
\name{renderQueryPhraseExactSynonym}
\alias{renderQueryPhraseExactSynonym}
\title{Render Query Concept Synonym Table}
\usage{
renderQueryPhraseExactSynonym(
  schema,
  caseInsensitive = TRUE,
  phrase,
  limit_n = NULL
)
}
\description{
Render Query Concept Synonym Table
}
