% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_functions.R
\name{renderQueryPhraseExactSynonym}
\alias{renderQueryPhraseExactSynonym}
\title{Render Query Concept Synonym Table}
\usage{
renderQueryPhraseExactSynonym(schema, phrase, caseInsensitive = TRUE)
}
\description{
Render Query Concept Synonym Table
}
