% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprec-render.R
\name{renderQueryPhraseLikeSynonym}
\alias{renderQueryPhraseLikeSynonym}
\title{Render Query Concept Synonym Table}
\usage{
renderQueryPhraseLikeSynonym(vocab_schema, phrase, caseInsensitive = TRUE)
}
\description{
Render Query Concept Synonym Table
}
