% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{renderSynonyms}
\alias{renderSynonyms}
\title{Render SQL Statement for Concept Synonyms}
\usage{
renderSynonyms(concept_id, vocab_schema = NULL, language_concept_id = 4180186)
}
\arguments{
\item{vocab_schema}{If NULL, defaults to the public schema.}
}
\description{
Get all the synonyms for a concept by concept_id.
}
