% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_ancestor_matrix.R
\name{return_ancestor_matrix}
\alias{return_ancestor_matrix}
\title{Return a Ancestor Matrix}
\usage{
return_ancestor_matrix(descendant_concept_ids)
}
\description{
This function returns a matrix of ancestors by descendants to view if a group of concepts have any common ancestors
}
\examples{
# Random immunosuppressant concept ids
immunosuppressant_concept_ids <- c("35807335","35807331", "21603616", "21600651", "21605199", "21602723") 
return_ancestor_matrix(immunosuppressant_concept_ids)
}
