% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.R
\name{sendAthena}
\alias{sendAthena}
\title{Send a SQL Statement}
\usage{
sendAthena(
  conn,
  conn_fun = "connectAthena()",
  sql_statement,
  verbose = TRUE,
  render_sql = TRUE,
  render_only = FALSE
)
}
\arguments{
\item{conn}{PARAM_DESCRIPTION, Default: NULL}

\item{sql_statement}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Unlike the \code{\link{queryAthena}} function, the render_sql parameter provides a pause after rendering in case the user wants to copy and paste the rendered SQL into a client if the session is interactive. This is particularly useful with large operations that are better executed within a background client.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[pg13]{send}}
}
