% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{unbox_label}
\alias{unbox_label}
\title{Unbox Strip}
\usage{
unbox_label(data, label_col, row_sep = "\\n", remove = FALSE)
}
\description{
First separates rows by the `row_sep` argument, followed by unmerging the strip
}
\seealso{
\code{\link[tidyr]{separate_rows}}
}
