% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{unbox_strip}
\alias{unbox_strip}
\title{Unbox Strip}
\usage{
unbox_strip(
  data,
  strip_col,
  sep = "\\n",
  suffix = NULL,
  prefix = NULL,
  remove = TRUE,
  r_trimws = TRUE
)
}
\description{
Unboxing refers to the process of both unmerging a column, either a label or a strip,
while also separating rowwise by the `sep` argument. It is meant to isolate
the Concept attributes from a merge column while also taking in account that
rowwise aggregates may have also occurred, and without separating these values out,
some concepts in the source data may be misparsed.
}
\examples{
library(tidyverse)
library(chariot)

get_strip(concept_id = 1112807,
         vocab_schema = "omop_vocabulary")

test_strip <- tibble::tibble(concept = "[V] [S] 1112807 aspirin [RxNorm 1191] [Drug] [Ingredient]")
unmerge_strip(test_strip,
              strip_col = concept)

unbox_strip(test_strip,
            strip_col = concept)

test_strip <- tibble::tibble(concept = "[V] [S] 1112807 aspirin [RxNorm 1191] [Drug] [Ingredient]\n[V] [S] 1112807 aspirin [RxNorm 1191] [Drug] [Ingredient]")
unbox_strip(test_strip,
            strip_col = concept)

}
\seealso{
\code{\link[tidyr]{separate_rows}}

Other concept format functions: 
\code{\link{filter_at_all_strip}()},
\code{\link{filter_at_any_strip}()},
\code{\link{filter_strip}()},
\code{\link{get_strip}()},
\code{\link{label_to_strip}()},
\code{\link{merge_label}()},
\code{\link{merge_strip}()},
\code{\link{unbox_label}()},
\code{\link{unmerge_label}()},
\code{\link{unmerge_strip}()}

Other unboxing functions: 
\code{\link{unbox_label}()}
}
\concept{concept format functions}
\concept{unboxing functions}
