% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{unbox_strip}
\alias{unbox_strip}
\title{Unbox Strip}
\usage{
unbox_strip(
  data,
  strip_col,
  sep = "\\n",
  suffix = NULL,
  prefix = NULL,
  remove = TRUE,
  r_trimws = TRUE
)
}
\description{
First separates rows by the `row_sep` argument, followed by unmerging the strip
}
\examples{
library(tidyverse)
library(chariot)

get_strip(concept_id = 1112807,
         vocab_schema = "omop_vocabulary")

test_strip <- tibble::tibble(concept = "[V] [S] 1112807 aspirin [RxNorm 1191] [Drug] [Ingredient]")
unmerge_strip(test_strip,
              strip_col = concept)

unbox_strip(test_strip,
            strip_col = concept)

test_strip <- tibble::tibble(concept = "[V] [S] 1112807 aspirin [RxNorm 1191] [Drug] [Ingredient]\n[V] [S] 1112807 aspirin [RxNorm 1191] [Drug] [Ingredient]")
unbox_strip(test_strip,
            strip_col = concept)

}
\seealso{
\code{\link[tidyr]{separate_rows}}
}
