% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{unmerge_label}
\alias{unmerge_label}
\title{Parse a Concept Label}
\usage{
unmerge_label(data, label_col, remove = FALSE)
}
\description{
Parse a concept Label in the format of "{concept_id} {concept_name}".
}
\seealso{
\code{\link[tidyr]{extract}}
}
